### 2023-04-21 696a855
Updated package requirements. Fixed deletion of user group from user not applied (and added related test).
