### 2023-05-02 712ce0d
Refs #209. Added "project_enabled" and "project_description" fields, and adjusted logic, APIs and tests.
