### 2023-05-03 2da5f9f
Refs #211 Decorator to verify global roles for a user in a service
