### 2023-05-03 06d82c7
Refs #211 Decorator to verify global roles for a user in a service
