### 2023-05-03 5255502
Refs #211 Decorator to verify global roles for a user in a service, decorator order is important.
