+++ teraserver/python/services/LoggingService/API/QueryLoginEntries.py
@@ -42,6 +42,7 @@ class QueryLoginEntries(Resource):
                         501: 'Not implemented.',
                         403: 'Logged user doesn\'t have permission to access the requested data'})
     @ServiceAccessManager.token_required(allow_dynamic_tokens=True, allow_static_tokens=False)
+    @ServiceAccessManager.service_user_roles_required(['admin'])
     def get(self):
 
         args = get_parser.parse_args()
