### 2023-05-05 f1b12d1
Fixed participant login with token issue with package upgrade.
