### 2023-05-09 c861bd8
Refs #211, cleanup confusing comments
