+++ teraserver/python/opentera/services/TeraParticipantClient.py
@@ -12,15 +12,8 @@ class TeraParticipantClient:
         self.__id_participant = token_dict['id_participant']
         self.__participant_token = token
 
-        # A little trick here to get the right URL for the server if we are using a proxy
         backend_hostname = config_man.backend_config["hostname"]
         backend_port = str(config_man.backend_config["port"])
-
-        #if 'X-Externalhost' in request.headers:
-        #    backend_hostname = request.headers['X-Externalhost']
-
-        #if 'X-Externalport' in request.headers:
-        #    backend_port = request.headers['X-Externalport']
 
         self.__backend_url = 'https://' + backend_hostname + ':' + backend_port
         self.__config_man = config_man
