+++ teraserver/python/opentera/services/TeraUserClient.py
@@ -12,15 +12,8 @@ class TeraUserClient:
         self.__user_token = token
         self.__user_superadmin = token_dict['user_superadmin']
 
-        # A little trick here to get the right URL for the server if we are using a proxy
         backend_hostname = config_man.backend_config["hostname"]
         backend_port = str(config_man.backend_config["port"])
-
-        # if 'X-Externalhost' in request.headers:
-        #   backend_hostname = request.headers['X-Externalhost']
-
-        #if 'X-Externalport' in request.headers:
-        #    backend_port = request.headers['X-Externalport']
 
         self.__backend_url = 'https://' + backend_hostname + ':' + backend_port
 
