### 2023-05-10 7054998
Refs #209. Prevented enabling participants part of a disabled project.
