+++ teraserver/python/opentera/forms/TeraDeviceForm.py
@@ -33,7 +33,7 @@ class TeraDeviceForm:
         section.add_item(TeraFormItem("id_device_subtype", gettext("Device Sub-Type"), "array", item_required=False,
                                       item_condition=TeraFormItemCondition("id_device_type", "=", "changed",
                                                                            "/api/user/devicesubtypes?id_device_type=")))
-        section.add_item(TeraFormItem("device_token", gettext("Access Token"), "label",
+        section.add_item(TeraFormItem("device_token", gettext("Access Token"), "longlabel",
                                       item_options={"readonly": False}))
         section.add_item(TeraFormItem("device_certificate", gettext("Certificate"), "hidden",
                                       item_options={"readonly": True}))
