### 2023-05-11 ac69cb5
Refs #213. Added UserQueryServiceRole API and related tests.
