### 2023-05-11 33e0285
Refs #213. Added 'id_user' parameter to UserQueryServiceAccess
