### 2023-05-12 b180cbd
Refs #210. Added "services/roles" API.
