### 2023-05-15 92af5ae
Refs #210. Added ServiceQueryProjects POST and DELETE API
