+++ teraserver/python/tests/modules/FlaskModule/API/service/test_ServiceQueryProjects.py
@@ -1,6 +1,7 @@
 from BaseServiceAPITest import BaseServiceAPITest
 from opentera.db.models.TeraProject import TeraProject
 from opentera.db.models.TeraSite import TeraSite
+from opentera.db.models.TeraServiceProject import TeraServiceProject
 
 
 class ServiceQueryProjectsTest(BaseServiceAPITest):
@@ -75,3 +76,86 @@ class ServiceQueryProjectsTest(BaseServiceAPITest):
             for json_project in response.json:
                 project: TeraProject = TeraProject.get_project_by_id(json_project['id_project'])
                 self.assertEqual(project.to_json(minimal=True), json_project)
+
+    def test_post_and_delete_endpoint_with_token(self):
+        with self._flask_app.app_context():
+            # Post with missing informations
+            json_data = {
+                'project_name': 'Testing123',
+            }
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(400, response.status_code, msg="Missing project struct")
+
+            json_data = {
+                'project': {
+                    'project_name': 'Testing123'
+                }
+            }
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(400, response.status_code, msg="Missing id_project")
+
+            json_data['project']['id_project'] = 0
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(400, response.status_code, msg="Missing id_site")
+
+            # Post in a site where service isn't associated
+            json_data['project']['id_site'] = 2
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(403, response.status_code, msg="No access to site")
+
+            # All good now
+            json_data['project']['id_site'] = 1
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(200, response.status_code, msg="New OK")
+            project_data = response.json
+            project_id = project_data['id_project']
+
+            # Check that service was correctly associated to that new project
+            sp: TeraServiceProject = TeraServiceProject.get_service_project_for_service_project(project_id=project_id,
+                                                                                                service_id=
+                                                                                                self.id_service)
+            self.assertIsNotNone(sp)
+
+            # Post update to project without association to service
+            json_data = {
+                'project': {
+                    'id_project': 3,
+                    'id_site': 2,
+                    'project_name': 'New Project Name'
+                }
+            }
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(403, response.status_code, msg="No access to project")
+
+            json_data = {
+                'project': {
+                    'id_project': project_id,
+                    'id_site': 2,
+                    'project_name': 'New Project Name',
+                    'invalid_parameter': -1
+                }
+            }
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(403, response.status_code, msg="No access to site")
+
+            del json_data['project']['id_site']
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(500, response.status_code, msg="Invalid parameter")
+
+            del json_data['project']['invalid_parameter']
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(200, response.status_code, msg="Update OK")
+
+            # Delete denied (not associated to service)
+            response = self._delete_with_service_token_auth(self.test_client, token=self.service_token,
+                                                            params={'id': 3})
[…]
