### 2023-05-16 fb922be
Refs #212, Disabling tokens when logout, use redis to store disabled tokens.
