### 2023-05-16 1b0caf9
Refs #212, Disabling tokens when logout, use redis to store disabled tokens.
