### 2023-05-17 a74a790
Refs #210. Added ServiceQueryUserGroups API.
