+++ teraserver/python/tests/services/LoggingService/BaseLoggingServiceAPITest.py
@@ -8,6 +8,7 @@ import random
 from string import digits, ascii_lowercase, ascii_uppercase
 from FakeLoggingService import FakeLoggingService
 from opentera.services.ServiceAccessManager import ServiceAccessManager
+from requests.auth import _basic_auth_str
 
 
 class BaseLoggingServiceAPITest(unittest.TestCase):
@@ -18,6 +19,8 @@ class BaseLoggingServiceAPITest(unittest.TestCase):
     participant_token_key = ''.join(random.choice(digits + ascii_lowercase + ascii_uppercase) for _ in range(36))
     service_token_key = ''.join(random.choice(digits + ascii_lowercase + ascii_uppercase) for _ in range(36))
     device_token_key = ''.join(random.choice(digits + ascii_lowercase + ascii_uppercase) for _ in range(36))
+    user_login_endpoint = '/api/user/login'
+    user_logout_endpoint = '/api/user/logout'
 
     @classmethod
     def setUpClass(cls):
@@ -82,6 +85,24 @@ class BaseLoggingServiceAPITest(unittest.TestCase):
         else:
             headers = {}
 
+        return client.get(endpoint, headers=headers, query_string=params)
+
+    def _get_with_user_token_auth(self, client: FlaskClient, token: str = '', params=None, endpoint=None):
+        if params is None:
+            params = {}
+        if endpoint is None:
+            endpoint = self.test_endpoint
+        headers = {'Authorization': 'OpenTera ' + token}
+        return client.get(endpoint, headers=headers, query_string=params)
+
+    def _get_with_user_http_auth(self, client: FlaskClient, username: str = '', password: str = '',
+                                 params=None, endpoint=None):
+        if params is None:
+            params = {}
+        if endpoint is None:
+            endpoint = self.test_endpoint
+
+        headers = {'Authorization': _basic_auth_str(username, password)}
         return client.get(endpoint, headers=headers, query_string=params)
 
     def _post_with_service_token_auth(self, client: FlaskClient, token: str = '', json: dict = None,
