+++ teraserver/python/env/requirements.txt
@@ -7,30 +7,30 @@ SQLAlchemy==1.4.46
 sqlalchemy-schemadisplay==1.3
 pydot==1.4.2
 psycopg2-binary==2.9.6
-Flask==2.2.3
+Flask==2.3.2
 Flask-SQLAlchemy==3.0.3
 Flask-Login==0.6.2
 Flask-Login-Multi==0.1.2
 Flask-HTTPAuth==4.8.0
-Flask-SocketIO==5.3.3
-Flask-Session==0.4.1
+Flask-SocketIO==5.3.4
+Flask-Session==0.5.0
 flask-restx==1.1.0
 Flask-Security==3.0.0
 Flask-Babel==3.1.0
 Flask-BabelEx==0.9.4
 Flask-Migrate==4.0.4
 flask-swagger-ui==4.11.1
-Flask-Limiter==3.3.0
+Flask-Limiter==3.3.1
 Flask-Mail==0.9.1
 Flask-Principal==0.4.0
-redis==4.5.4
+redis==4.5.5
 txredisapi==1.4.9
 passlib==1.7.4
 bcrypt==4.0.1
 WTForms==3.0.1
 pyOpenSSL==23.1.1
 service-identity==21.1.0
-PyJWT==2.6.0
+PyJWT==2.7.0
 pylzma==0.5.0
 bz2file==0.98
 python-slugify==8.0.1
