### 2023-05-18 fc36280
Refs #182, Disabling or deleting users, participants, devices will disconnect.
