+++ teraserver/python/modules/FlaskModule/FlaskModule.py
@@ -4,12 +4,10 @@ from flask_restx import Api, Namespace
 from opentera.config.ConfigManager import ConfigManager
 from flask_babel import Babel
 from opentera.modules.BaseModule import BaseModule, ModuleNames
-from opentera.modules.BaseModule import create_module_event_topic_from_name, create_module_message_topic_from_name
 from opentera.db.models.TeraServerSettings import TeraServerSettings
 from opentera.OpenTeraServerVersion import opentera_server_version_string
 import redis
 from modules.Globals import opentera_doc_url
-import opentera.messages.python as messages
 
 
 # Flask application
@@ -126,27 +124,6 @@ class FlaskModule(BaseModule):
         """
         print('FlaskModule - Received message ', pattern, channel, message)
         pass
-
-    def _send_disconnect_module_message(self, uuid: str):
-        module_message = self.create_module_message(
-            src=create_module_message_topic_from_name(ModuleNames.FLASK_MODULE_NAME),
-            dest=create_module_event_topic_from_name(ModuleNames.TWISTED_MODULE_NAME, uuid)
-        )
-        server_command = messages.ServerCommand()
-        server_command.type = messages.ServerCommand.CMD_CLIENT_DISCONNECT
-        command_message = messages.Any()
-        command_message.Pack(server_command)
-        module_message.data.extend([command_message])
-        self.send_module_message(module_message)
-
-    def send_user_disconnect_module_message(self, user_uuid: str):
-        self._send_disconnect_module_message(user_uuid)
-
-    def send_participant_disconnect_module_message(self, participant_uuid: str):
-        self._send_disconnect_module_message(participant_uuid)
-
-    def send_device_disconnect_module_message(self, device_uuid: str):
-        self._send_disconnect_module_message(device_uuid)
 
     @staticmethod
     def init_user_api(module: object, namespace: Namespace, additional_args: dict = dict()):
