+++ teraserver/python/opentera/db/models/TeraSessionEvent.py
@@ -40,8 +40,7 @@ class TeraSessionEvent(BaseModel, SoftDeleteMixin):
         CUSTOM_EVENT10 = 109
 
     __tablename__ = 't_sessions_events'
-    id_session_event = Column(Integer, Sequence('id_session_events_sequence'), primary_key=True,
-                                 autoincrement=True)
+    id_session_event = Column(Integer, Sequence('id_session_events_sequence'), primary_key=True, autoincrement=True)
     id_session = Column(Integer, ForeignKey('t_sessions.id_session', ondelete='cascade'), nullable=False)
     id_session_event_type = Column(Integer, nullable=False)
     session_event_datetime = Column(TIMESTAMP(timezone=True), nullable=False)
