+++ teraserver/python/env/requirements.txt
@@ -3,7 +3,7 @@ Twisted==22.10.0
 treq==22.2.0
 cryptography==40.0.2
 autobahn==23.1.2
-SQLAlchemy==1.4.46
+SQLAlchemy==1.4.48
 sqlalchemy-schemadisplay==1.3
 pydot==1.4.2
 psycopg2-binary==2.9.6
@@ -34,7 +34,7 @@ PyJWT==2.7.0
 pylzma==0.5.0
 bz2file==0.98
 python-slugify==8.0.1
-websocket-client==1.5.1
+websocket-client==1.5.2
 pytest==7.3.1
 jsonschema==4.17.3
 Jinja2==3.1.2
