### 2023-05-23 544f7ba
Refs #214, service is now updated in redis when modified.
