+++ teraserver/python/env/CMakeLists.txt
@@ -8,7 +8,7 @@ file(MAKE_DIRECTORY ${ENV_PATH})
 
 # Define packages to install
 set (conda_packages
-    protobuf
+    protobuf==3.20.3
 )
 
 if(WIN32)
