+++ teraserver/python/setup.py
@@ -9,7 +9,7 @@ with open("env/requirements.txt", "r", encoding="utf-8") as f:
 
 setuptools.setup(
     name="opentera",
-    version="1.2.2",
+    version="1.2.3",
     author="Dominic Létourneau, Simon Brière",
     author_email="dominic.letourneau@usherbrooke.ca, simon.briere@usherbrooke.ca",
     description="OpenTera base package",
