+++ docs/conf.py
@@ -26,7 +26,8 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'venv']
 
 html_theme = 'sphinx_rtd_theme'
 html_theme_options = {
-    'navigation_depth': -1
+    'navigation_depth': -1,
+    'display_version': True
 }
 
 source_suffix = {
