### 2023-05-30 #5497 @editorialbot
[REVIEW]: OpenTera: A Framework for Telehealth Applications
**Submitting author:** @doumdi (<a href="http://orcid.org/0000-0001-7825-7533">Dominic Létourneau</a>)
**Repository:** https://github.com/introlab/opentera
**Branch with paper.md** (empty if default branch): main
**Version:** v1.2.4-zenodo
