### 2023-05-30 85a782d
Refs #216. Added POST and DELETE for ServiceQueryTestTypes
