+++ teraserver/python/tests/modules/FlaskModule/API/service/test_ServiceQueryTestTypes.py
@@ -1,11 +1,12 @@
 from typing import List
 from BaseServiceAPITest import BaseServiceAPITest
-from modules.FlaskModule.FlaskModule import flask_app
 from opentera.db.models.TeraTestType import TeraTestType
 from opentera.db.models.TeraTestTypeSite import TeraTestTypeSite
 from opentera.db.models.TeraTestTypeProject import TeraTestTypeProject
 from opentera.db.models.TeraService import TeraService
 from opentera.db.models.TeraParticipant import TeraParticipant
+from opentera.db.models.TeraTest import TeraTest
+import datetime
 
 
 class ServiceQueryTestTypesTest(BaseServiceAPITest):
@@ -20,6 +21,11 @@ class ServiceQueryTestTypesTest(BaseServiceAPITest):
     def test_get_endpoint_no_auth(self):
         with self._flask_app.app_context():
             response = self.test_client.get(self.test_endpoint)
+            self.assertEqual(401, response.status_code)
+
+    def test_post_endpoint_no_auth(self):
+        with self._flask_app.app_context():
+            response = self.test_client.post(self.test_endpoint, query_string=None, headers=None, json={})
             self.assertEqual(401, response.status_code)
 
     def test_get_endpoint_with_token_auth_no_params(self):
@@ -122,3 +128,90 @@ class ServiceQueryTestTypesTest(BaseServiceAPITest):
             self.assertEqual(1, len(response.json))
             self.assertTrue(response.json[0]['id_test_type'] in accessible_types)
             self.assertEqual(response.json[0]['id_test_type'], test_types[0].id_test_type)
+
+    def test_post_and_delete(self):
+        with self._flask_app.app_context():
+            # New with minimal infos
+            json_data = {
+                'test_type': {
+                    'test_type_name': 'Test'
+                }
+            }
+
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(400, response.status_code, msg="Missing id_test_type")
+
+            json_data['test_type']['id_test_type'] = 0
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(200, response.status_code, msg="New OK")
+
+            json_data = response.json
+            self._checkJson(json_data)
+            current_id = json_data['id_test_type']
+            current_uuid = json_data['test_type_uuid']
+
+            json_data = {
+                'test_type': {
+                    'id_test_type': 3,
+                    'test_type_name': 'Test 2',
+                    'test_type_has_json_format': True
+                }
+            }
+
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(403, response.status_code, msg="No access to test type")
+
+            json_data['test_type']['id_test_type'] = current_id
+            json_data['test_type']['id_service'] = 1
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(403, response.status_code, msg="Not current service")
+
+            del json_data['test_type']['id_service']
+            response = self._post_with_service_token_auth(self.test_client, token=self.service_token, json=json_data)
+            self.assertEqual(200, response.status_code, msg="Update OK")
+
+            reply_data = response.json
+            self._checkJson(reply_data)
+            self.assertEqual(reply_data['test_type_name'], 'Test 2')
+            self.assertEqual(reply_data['test_type_has_json_format'], True)
+
+            response = self._delete_with_service_token_auth(self.test_client, token=self.service_token,
+                                                            params={'uuid': '0000000000000000'})
+            self.assertEqual(400, response.status_code, msg="Bad parameter")
+
[…]
