+++ teraserver/python/modules/FlaskModule/API/service/ServiceQueryTestTypes.py
@@ -1,5 +1,7 @@
 from flask_restx import Resource
 from flask_babel import gettext
+from flask import request
+from sqlalchemy import exc
 from modules.LoginModule.LoginModule import LoginModule, current_service
 from modules.FlaskModule.FlaskModule import service_api_ns as api
 from modules.DatabaseModule.DBManager import DBManager
@@ -16,6 +18,13 @@ get_parser.add_argument('id_project', type=int, help='ID of the project to query
 get_parser.add_argument('id_participant', type=int, help='ID of the participant to query types for')
 get_parser.add_argument('test_type_key', type=str, help='Test type key to query for')
 get_parser.add_argument('id_test_type', type=int, help='ID of the test type to query for')
+
+post_parser = api.parser()
+post_schema = api.schema_model('service_test_type', {'properties': TeraTestType.get_json_schema(), 'type': 'object',
+                                                     'location': 'json'})
+
+delete_parser = api.parser()
+delete_parser.add_argument('uuid', type=str, help='Test type UUID to delete', required=True)
 
 
 class ServiceQueryTestTypes(Resource):
@@ -62,3 +71,98 @@ class ServiceQueryTestTypes(Resource):
             test_types = service_access.get_accessible_tests_types()
 
         return [tt.to_json() for tt in test_types]
+
+    @api.doc(description='Create / update test type. id_test_type must be set to "0" to create a new '
+                         'type.',
+             responses={200: 'Success',
+                        403: 'Service can\'t create/update the specified test type',
+                        400: 'Badly formed JSON or missing field in the JSON body',
+                        500: 'Internal error when saving test type'},
+             params={'token': 'Secret token'})
+    @api.expect(post_schema)
+    @LoginModule.service_token_or_certificate_required
+    def post(self):
+        # Using request.json instead of parser, since parser messes up the json!
+        if 'test_type' not in request.json:
+            return gettext('Missing test_type'), 400
+
+        json_test_type = request.json['test_type']
+
+        # Validate if we have an id
+        if 'id_test_type' not in json_test_type:
+            return gettext('Missing id_test_type'), 400
+
+        if 'id_service' in json_test_type and json_test_type['id_service'] != current_service.id_service:
+            return gettext('Forbidden'), 403
+
+        if json_test_type['id_test_type'] != 0:
+            test_type = TeraTestType.get_test_type_by_id(json_test_type['id_test_type'])
+            if not test_type or test_type.id_service != current_service.id_service:
+                return gettext('Forbidden'), 403
+            # Updating
+            try:
+                TeraTestType.update(json_test_type['id_test_type'], json_test_type)
+            except exc.SQLAlchemyError as e:
+                import sys
+                print(sys.exc_info())
+                self.module.logger.log_error(self.module.module_name,
+                                             ServiceQueryTestTypes.__name__,
+                                             'post', 500, 'Database error', str(e))
+                return gettext('Database error'), 500
+        else:
+            # New
+            # Always force id to this service
+            json_test_type['id_service'] = current_service.id_service
+            try:
+                missing_fields = TeraTestType.validate_required_fields(json_data=json_test_type)
+                if missing_fields:
+                    return gettext('Missing fields') + ': ' + str([field for field in missing_fields]), 400
+
+                new_tt = TeraTestType()
+                new_tt.from_json(json_test_type)
+                TeraTestType.insert(new_tt)
+                # Update ID for further use
+                json_test_type['id_test_type'] = new_tt.id_test_type
+            except exc.SQLAlchemyError as e:
+                import sys
+                print(sys.exc_info())
+                self.module.logger.log_error(self.module.module_name,
+                                             ServiceQueryTestTypes.__name__,
[…]
