### 2023-05-30 ed092d4
Refs #210. Added ServiceQueryTestTypeProjects Added "session_type_project" and "session_type_site" in database update message.
