+++ teraserver/python/modules/FlaskModule/API/service/ServiceQueryTestTypeProjects.py
@@ -1,0 +1,271 @@
+from flask import request
+from flask_restx import Resource, reqparse, inputs
+from modules.LoginModule.LoginModule import current_service, LoginModule
+from modules.FlaskModule.FlaskModule import user_api_ns as api
+from opentera.db.models.TeraTestTypeProject import TeraTestTypeProject
+from opentera.db.models.TeraTestTypeSite import TeraTestTypeSite
+from opentera.db.models.TeraTestType import TeraTestType
+from opentera.db.models.TeraProject import TeraProject
+from modules.DatabaseModule.DBManager import DBManager
+from sqlalchemy.exc import InvalidRequestError
+from sqlalchemy import exc, inspect
+from flask_babel import gettext
+
+# Parser definition(s)
+get_parser = api.parser()
+get_parser.add_argument('id_project', type=int, help='Project ID to query associated test types from')
+get_parser.add_argument('id_test_type', type=int, help='Test type ID to query associated projects from')
+
+get_parser.add_argument('with_sites', type=inputs.boolean, help='Used with id_test_type. Also return site '
+                                                                'information of the returned projects.')
+
+post_parser = api.parser()
+post_schema = api.schema_model('service_test_type_project', {'properties': TeraTestTypeProject.get_json_schema(),
+                                                             'type': 'object',
+                                                             'location': 'json'})
+
+delete_parser = reqparse.RequestParser()
+delete_parser.add_argument('id', type=int, help='Specific test type - project association ID to delete. '
+                                                'Be careful: this is not the test-type or project ID, but the ID'
+                                                ' of the association itself!', required=True)
+
+
+class ServiceQueryTestTypeProjects(Resource):
+
+    def __init__(self, _api, *args, **kwargs):
+        Resource.__init__(self, _api, *args, **kwargs)
+        self.module = kwargs.get('flaskModule', None)
+        self.test = kwargs.get('test', False)
+
+    @api.doc(description='Get test types that are associated with a project. Only one "ID" parameter required and '
+                         'supported at once.',
+             responses={200: 'Success - returns list of test-types - projects association',
+                        400: 'Required parameter is missing (must have at least one id)',
+                        500: 'Error when getting association'},
+             params={'token': 'Secret token'})
+    @api.expect(get_parser)
+    @LoginModule.service_token_or_certificate_required
+    def get(self):
+        service_access = DBManager.serviceAccess(current_service)
+        args = get_parser.parse_args()
+
+        test_type_projects = []
+        # If we have no arguments, return error
+        if not any(args.values()):
+            return gettext('Missing arguments'), 400
+
+        if args['id_project']:
+            if args['id_project'] in service_access.get_accessible_projects_ids():
+                test_type_projects = TeraTestTypeProject.get_tests_types_for_project(args['id_project'])
+        elif args['id_test_type']:
+            if args['id_test_type'] in service_access.get_accessible_tests_types_ids():
+                test_type_projects = TeraTestTypeProject.get_projects_for_test_type(args['id_test_type'])
+        try:
+            ttp_list = []
+            for ttp in test_type_projects:
+                json_ttp = ttp.to_json()
+                obj_type = inspect(ttp)
+                if not obj_type.transient:
+                    json_ttp['test_type_name'] = ttp.test_type_project_test_type.test_type_name
+                    json_ttp['project_name'] = ttp.test_type_project_project.project_name
+                    if args['with_sites']:
+                        json_ttp['id_site'] = ttp.test_type_project_project.id_site
+                        json_ttp['site_name'] = ttp.test_type_project_project.project_site.site_name
+                else:
+                    # Temporary object, a not-committed object, result of listing projects not associated to a
+                    # test type.
[…]
