+++ teraserver/python/opentera/db/models/TeraSessionTypeProject.py
@@ -1,5 +1,6 @@
 from opentera.db.Base import BaseModel
 from opentera.db.models.TeraSession import TeraSession
+from opentera.db.models.TeraSessionType import TeraSessionType
 from opentera.db.SoftDeleteMixin import SoftDeleteMixin
 from opentera.db.SoftInsertMixin import SoftInsertMixin
 from sqlalchemy import Column, ForeignKey, Integer, Sequence
@@ -28,6 +29,23 @@ class TeraSessionTypeProject(BaseModel, SoftDeleteMixin, SoftInsertMixin):
         rval = super().to_json(ignore_fields=ignore_fields)
 
         return rval
+
+    def to_json_create_event(self):
+        json_data = self.to_json(minimal=True)
+        # Query session type information
+        session_type = TeraSessionType.get_session_type_by_id(self.id_session_type)
+        if session_type:
+            json_data['session_type_id_service'] = session_type.id_service
+        return json_data
+
+    def to_json_update_event(self):
+        json_data = self.to_json(minimal=True)
+        json_data['session_type_id_service'] = self.session_type_project_session_type.id_service
+        return json_data
+
+    def to_json_delete_event(self):
+        # Minimal information, delete can not be filtered
+        return {'id_session_type_project': self.id_session_type_project}
 
     @staticmethod
     def create_defaults(test=False):

[…]
