### 2023-05-30 1418360
Refs #210. Added automatic association of test type to site when using ServiceQueryTestTypeProjects API. Added "id_session_type" parameter in ServiceQuerySessionTypes API.
