+++ teraserver/python/modules/FlaskModule/API/service/ServiceQuerySessionTypes.py
@@ -1,10 +1,10 @@
 from flask_restx import Resource
-from flask_babel import gettext
 from modules.LoginModule.LoginModule import LoginModule, current_service
 from modules.FlaskModule.FlaskModule import service_api_ns as api
 from modules.DatabaseModule.DBManager import DBManager
 from opentera.db.models.TeraSessionTypeProject import TeraSessionTypeProject
 from opentera.db.models.TeraSessionTypeSite import TeraSessionTypeSite
+from opentera.db.models.TeraSessionType import TeraSessionType
 from opentera.db.models.TeraParticipant import TeraParticipant
 
 # Parser definition(s)
@@ -12,6 +12,7 @@ get_parser = api.parser()
 get_parser.add_argument('id_site', type=int, help='ID of the site to query session types for')
 get_parser.add_argument('id_project', type=int, help='ID of the project to query session types for')
 get_parser.add_argument('id_participant', type=int, help='ID of the participant to query types for')
+get_parser.add_argument('id_session_type', type=int, help='ID of the session type to query')
 
 
 class ServiceQuerySessionTypes(Resource):
@@ -47,6 +48,9 @@ class ServiceQuerySessionTypes(Resource):
                 part_info = TeraParticipant.get_participant_by_id(args['id_participant'])
                 session_types = [st.session_type_project_session_type for st in
                                  TeraSessionTypeProject.get_sessions_types_for_project(part_info.id_project)]
+        elif args['id_session_type']:
+            if args['id_session_type'] in service_access.get_accessible_sessions_types_ids():
+                session_types = [TeraSessionType.get_session_type_by_id(args['id_session_type'])]
         else:
             session_types = service_access.get_accessible_sessions_types()
 
