### 2023-05-30 6a1a40a
Fixed failed test in ServiceQueryTestTypeProjectTest
