+++ teraserver/CMakeLists.txt
@@ -10,7 +10,7 @@ endif(NOT CMAKE_BUILD_TYPE)
 # Software version
 SET(OPENTERA_VERSION_MAJOR "1")
 SET(OPENTERA_VERSION_MINOR "2")
-SET(OPENTERA_VERSION_PATCH "3")
+SET(OPENTERA_VERSION_PATCH "4")
 
 SET(OPENTERA_SERVER_VERSION OpenTera_v${OPENTERA_VERSION_MAJOR}.${OPENTERA_VERSION_MINOR}.${OPENTERA_VERSION_PATCH})
 
