### 2023-06-12 90b0df8
Fixed issues in UserQueryTestType for inaccessible test types.
