### 2023-06-12 b1f375c
Refs #219, Fixed certificates error (not checking path) when not using ssl
