+++ teraserver/python/certificates/devices/README.TXT
@@ -1,1 +1,0 @@
-Device test certificates here...
