### 2023-06-12 2489d6a
Fixed DBManagerTeraUserAccess query_test_type
