+++ teraserver/python/modules/DatabaseModule/DBManagerTeraUserAccess.py
@@ -566,7 +566,8 @@ class DBManagerTeraUserAccess:
         # site_ids = self.get_accessible_sites_ids()
         # proj_ids = self.get_accessible_projects_ids()
         service_ids = self.get_accessible_services_ids()
-        test_type = TeraTestType.query.filter_by(id_test_type=test_type_id).first()
+        test_type = TeraTestType.query.filter_by(id_test_type=test_type_id)\
+            .filter(TeraTestType.id_service.in_(service_ids)).first()
         # .filter(TeraSite.id_site.in_(site_ids))\
         # .filter(TeraTestType.id_service.in_(service_ids)).filter(TeraProject.id_project.in_(proj_ids)).first()
         return test_type
