### 2023-06-22 0541229
Changed delete access check in UserQueryTestType
