### 2023-07-05 #5497 @doumdi
@galessiorob @Rocsg  I am not sure I understand the V2 vs original version. The [paper in the "main" branch](https://github.com/introlab/opentera/tree/main/joss-paper) did not change since our submission three months ago. Of course, development continues on OpenTera and we did merge some  bug fixes and Python packages upgrade fix in the main branch. Documentation / features did not change however. How can I be of assistance to make sure you have everything you need ?
