### 2023-07-05 #5497 @galessiorob
Thanks for chiming in @doumdi, I think (please correct me if wrong @Rocsg) the question came up due to [the previous publishing of the software in 2022 ](https://pubmed.ncbi.nlm.nih.gov/35223361/) and [JOSS's guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) that [ask to disclose if co-publishing.](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software) From my purview, as long as this is disclosed and the software is a significant contribution, which in this case it is, then we are good to review and publish this paper (see: "JOSS considers submissions for which the implementation of the software itself reflects a substantial scientific effort" in the guidelines.)

I will check with the EiC about this to confirm shortly.
