### 2023-07-06 #5497 @Rocsg
@galessiorob I think you understood exactly my point. @doumdi when I talk about "V2", I mean positioning the diff in OpenTera relatively to the version disclosed in the feb. 2022 paper in the health and Tech. journal. As a rule of thumb, I "measured" the activity of the github repo between the versions (since feb. 2022, and there is), and considered the diff between the journals in terms of scope, which can be substantial. From my experience, I published a paper in Bioinformatics which was NOT studied at all, and another one in JOSS which was really studied by the reviewers. So these publishing paths are significantly different, and complementary, to me. And I let this point to the editor representative, as I believe I am not in charge of that kind of questions in this process.

See you

Romain
