### 2023-07-06 7b10408
Refs #220, Allow participants and devices with static token to post file assets
