### 2023-07-06 9b5119c
Refs #221 Test for object is not None before handling update / delete
