### 2023-07-06 0ee5638
Refs #221 Test for object is not None before handling update / delete
