### 2023-07-07 #5497 @doumdi
@Rocsg, @galessiorob thanks for the clarifications. The focus of the two papers is different. The first paper aimed to describe specific applications related to COVID with a preliminary implementation of OpenTera. The current paper presents the implementation and design choices for a more generic OpenTera framework with a focus on open source implementation. Code quality, documentation, examples and usability have been greatly improved between these versions. There is little overlap between the two papers.
