+++ teraserver/python/env/requirements.txt
@@ -1,14 +1,14 @@
 pypiwin32==223; sys_platform == 'win32'
 Twisted==22.10.0
 treq==22.2.0
-cryptography==40.0.2
-autobahn==23.1.2
+cryptography==41.0.2
+autobahn==23.6.2
 SQLAlchemy==1.4.48
 sqlalchemy-schemadisplay==1.3
 pydot==1.4.2
 psycopg2-binary==2.9.6
 Flask==2.3.2
-Flask-SQLAlchemy==3.0.3
+Flask-SQLAlchemy==3.0.5
 Flask-Login==0.6.2
 Flask-Login-Multi==0.1.2
 Flask-HTTPAuth==4.8.0
@@ -23,19 +23,19 @@ flask-swagger-ui==4.11.1
 Flask-Limiter==3.3.1
 Flask-Mail==0.9.1
 Flask-Principal==0.4.0
-redis==4.5.5
-txredisapi==1.4.9
+redis==4.6.0
+txredisapi==1.4.10
 passlib==1.7.4
 bcrypt==4.0.1
 WTForms==3.0.1
-pyOpenSSL==23.1.1
-service-identity==21.1.0
+pyOpenSSL==23.2.0
+service-identity==23.1.0
 PyJWT==2.7.0
 pylzma==0.5.0
 bz2file==0.98
 python-slugify==8.0.1
-websocket-client==1.5.2
-pytest==7.3.1
-jsonschema==4.17.3
+websocket-client==1.6.1
+pytest==7.4.0
+jsonschema==4.18.0
 Jinja2==3.1.2
-ua-parser==0.16.1
+ua-parser==0.18.0
