+++ teraserver/python/env/requirements.txt
@@ -3,7 +3,7 @@ Twisted==22.10.0
 treq==22.2.0
 cryptography==41.0.2
 autobahn==23.6.2
-SQLAlchemy==1.4.48
+SQLAlchemy==2.0.18
 sqlalchemy-schemadisplay==1.3
 pydot==1.4.2
 psycopg2-binary==2.9.6
