+++ teraserver/python/modules/LoginModule/LoginModule.py
@@ -16,7 +16,7 @@ from opentera.services.DisabledTokenStorage import DisabledTokenStorage
 import datetime
 import redis
 
-from flask import request, _request_ctx_stack
+from flask import request, g
 from flask_babel import gettext
 from werkzeug.local import LocalProxy
 from flask_restx import reqparse
@@ -30,16 +30,16 @@ import modules.Globals as Globals
 from opentera.utils.UserAgentParser import UserAgentParser
 
 # Current participant identity, stacked
-current_participant = LocalProxy(lambda: getattr(_request_ctx_stack.top, 'current_participant', None))
+current_participant = LocalProxy(lambda: g.setdefault('current_participant', None))
 
 # Current device identity, stacked
-current_device = LocalProxy(lambda: getattr(_request_ctx_stack.top, 'current_device', None))
+current_device = LocalProxy(lambda: g.setdefault('current_device', None))
 
 # Current user identity, stacked
-current_user = LocalProxy(lambda: getattr(_request_ctx_stack.top, 'current_user', None))
+current_user = LocalProxy(lambda: g.setdefault('current_user', None))
 
 # Current service identity, stacked
-current_service = LocalProxy(lambda: getattr(_request_ctx_stack.top, 'current_service', None))
+current_service = LocalProxy(lambda: g.setdefault('current_service', None))
 
 # Authentication schemes for users
 user_http_auth = HTTPBasicAuth(realm='user')
@@ -102,7 +102,7 @@ class LoginModule(BaseModule):
 
         # We wait until we are connected to redis
         # Every 30 minutes?
-        loopDeferred = self.cleanup_disabled_tokens_loop_task.start(60.0 * 30)
+        self.cleanup_disabled_tokens_loop_task.start(60.0 * 30)
 
     def notify_module_messages(self, pattern, channel, message):
         """
@@ -171,8 +171,7 @@ class LoginModule(BaseModule):
             self.logger.send_login_event(sender='LoginModule.user_verify_password',
                                          level=messages.LogEvent.LOGLEVEL_ERROR,
                                          login_type=messages.LoginEvent.LOGIN_TYPE_PASSWORD,
-                                         login_status=
-                                         messages.LoginEvent.LOGIN_STATUS_FAILED_WITH_WRONG_USERNAME,
+                                         login_status=messages.LoginEvent.LOGIN_STATUS_FAILED_WITH_WRONG_USERNAME,
                                          client_name=login_infos['client_name'],
                                          client_version=login_infos['client_version'],
                                          client_ip=login_infos['client_ip'],
@@ -195,8 +194,7 @@ class LoginModule(BaseModule):
             self.logger.send_login_event(sender='LoginModule.user_verify_password',
                                          level=messages.LogEvent.LOGLEVEL_ERROR,
                                          login_type=messages.LoginEvent.LOGIN_TYPE_PASSWORD,
-                                         login_status=
-                                         messages.LoginEvent.LOGIN_STATUS_FAILED_WITH_MAX_ATTEMPTS_REACHED,
+                                         login_status=messages.LoginEvent.LOGIN_STATUS_FAILED_WITH_MAX_ATTEMPTS_REACHED,
                                          client_name=login_infos['client_name'],
                                          client_version=login_infos['client_version'],
                                          client_ip=login_infos['client_ip'],
@@ -209,7 +207,7 @@ class LoginModule(BaseModule):
         logged_user = TeraUser.verify_password(username=username, password=password, user=tentative_user)
 
         if logged_user and logged_user.is_active():
-            _request_ctx_stack.top.current_user = logged_user
+            g.current_user = logged_user
 
             # print('user_verify_password, found user: ', current_user)
             # current_user.update_last_online()
[…]
