+++ .github/workflows/gen-doc-and-deploy-to-github-pages.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.10.7]
+        python-version: [3.11.4]
 
     steps:
     - uses: actions/checkout@v3
