+++ .github/workflows/gen-doc-and-deploy-to-github-pages.yml
@@ -56,7 +56,7 @@ jobs:
       run: |
         cmake .
         make python-all
-        echo "OPENTERA_PYTHON=$(echo $PWD/python/env/python-3.10/bin/python)" >> $GITHUB_ENV
+        echo "OPENTERA_PYTHON=$(echo $PWD/python/env/python-3.11/bin/python)" >> $GITHUB_ENV
         echo "PYTHONPATH=$(echo $PWD/python)" >> $GITHUB_ENV
 
     - name: Generate Self Signed TLS Certificates
