### 2023-07-12 703faee
Refs #203. Implemented base for hard delete - working for assets and devices so far.
