+++ teraserver/python/opentera/db/SoftDeleteMixin.py
@@ -25,31 +25,35 @@ from opentera.db.SoftDeleteQueryRewriter import SoftDeleteQueryRewriter
 def activate_soft_delete_hook(deleted_field_name: str, disable_soft_delete_option_name: str):
     """Activate an event hook to rewrite the queries."""
     # Enable Soft Delete on all Relationship Loads which implement SoftDeleteMixin
-    # @listens_for(Session, "do_orm_execute")
-    # def soft_delete_execute(state: ORMExecuteState):
-    #     if not state.is_select:
-    #         return
-    #     if 'include_deleted' in state.session.info and len(state.session.info['include_deleted']) > 0:
-    #         print('test_include_deleted')
-    #         return
-    #
-    #     adapted = SoftDeleteQueryRewriter(deleted_field_name, disable_soft_delete_option_name).rewrite_statement(
-    #         state.statement
-    #     )
-    #     state.statement = adapted
-    @listens_for(Engine, "before_execute", retval=True)
-    def soft_delete_execute(conn: Connection, clauseelement, multiparams, params, execution_options):
-        if not isinstance(clauseelement, Select):
-            return clauseelement, multiparams, params
+    @listens_for(Session, "do_orm_execute")
+    def soft_delete_execute(state: ORMExecuteState):
+        if not state.is_select:
+            return
 
-        if disable_soft_delete_option_name in execution_options and execution_options[disable_soft_delete_option_name]:
-            # print('test_include_deleted')
-            return clauseelement, multiparams, params
+        if disable_soft_delete_option_name in state.execution_options \
+                and state.execution_options[disable_soft_delete_option_name]:
+            return
+
+        if 'include_deleted' in state.session.info and len(state.session.info['include_deleted']) > 0:
+            return
 
         adapted = SoftDeleteQueryRewriter(deleted_field_name, disable_soft_delete_option_name).rewrite_statement(
-            clauseelement
+            state.statement
         )
-        return adapted, multiparams, params
+        state.statement = adapted
+    # @listens_for(Engine, "before_execute", retval=True)
+    # def soft_delete_execute(conn: Connection, clauseelement, multiparams, params, execution_options):
+    #     if not isinstance(clauseelement, Select):
+    #         return clauseelement, multiparams, params
+    #
+    #     if disable_soft_delete_option_name in execution_options and execution_options[disable_soft_delete_option_name]:
+    #         # print('test_include_deleted')
+    #         return clauseelement, multiparams, params
+    #
+    #     adapted = SoftDeleteQueryRewriter(deleted_field_name, disable_soft_delete_option_name).rewrite_statement(
+    #         clauseelement
+    #     )
+    #     return adapted, multiparams, params
 
 
 def generate_soft_delete_mixin_class(
@@ -62,7 +66,9 @@ def generate_soft_delete_mixin_class(
     delete_method_default_value: Callable[[], Any] = lambda: datetime.utcnow(),
     generate_undelete_method: bool = True,
     undelete_method_name: str = "undelete",
-    handle_cascade_delete: bool = True
+    handle_cascade_delete: bool = True,
+    generate_hard_delete_method: bool = True,
+    hard_delete_method_name: str = "hard_delete"
 ) -> Type:
     """Generate the actual soft-delete Mixin class."""
     class_attributes = {deleted_field_name: Column(deleted_field_name, deleted_field_type)}
@@ -76,7 +82,6 @@ def generate_soft_delete_mixin_class(
     class_attributes['get_class_from_tablename'] = get_class_from_tablename
 
     if generate_delete_method:
-
         def delete_method(_self, v: Optional[Any] = None):
             setattr(_self, deleted_field_name, v or delete_method_default_value())
[…]
