+++ teraserver/python/tests/opentera/db/models/test_TeraDevice.py
@@ -1,5 +1,13 @@
+from opentera.db.models.TeraAsset import TeraAsset
+from opentera.db.models.TeraTest import TeraTest
+from opentera.db.models.TeraService import TeraService
 from opentera.db.models.TeraDevice import TeraDevice
 from opentera.db.models.TeraDeviceType import TeraDeviceType
+from opentera.db.models.TeraDeviceProject import TeraDeviceProject
+from opentera.db.models.TeraDeviceSite import TeraDeviceSite
+from opentera.db.models.TeraDeviceParticipant import TeraDeviceParticipant
+from opentera.db.models.TeraServiceConfig import TeraServiceConfig
+from opentera.db.models.TeraSession import TeraSession
 from tests.opentera.db.models.BaseModelsTest import BaseModelsTest
 
 
@@ -107,4 +115,115 @@ class TeraDeviceTest(BaseModelsTest):
             self.assertIsNotNone(device)
             self.assertIsNotNone(device.deleted_at)
 
+    def test_hard_delete(self):
+        with self._flask_app.app_context():
+            # Create a new device
+            device = TeraDevice()
+            device.device_name = 'Test device'
+            device.id_device_type = TeraDeviceType.get_device_type_by_key('capteur').id_device_type
+            TeraDevice.insert(device)
+            self.assertIsNotNone(device.id_device)
+            id_device = device.id_device
 
+            # Assign device to site
+            device_site = TeraDeviceSite()
+            device_site.id_site = 1
+            device_site.id_device = id_device
+            TeraDeviceSite.insert(device_site)
+            id_device_site = device_site.id_device_site
+
+            # Assign device to project
+            device_project = TeraDeviceProject()
+            device_project.id_device = id_device
+            device_project.id_project = 1
+            TeraDeviceProject.insert(device_project)
+            id_device_project = device_project.id_device_project
+
+            # Assign device to participants
+            device_participant = TeraDeviceParticipant()
+            device_participant.id_device = id_device
+            device_participant.id_participant = 1
+            TeraDeviceParticipant.insert(device_participant)
+            id_device_participant = device_participant.id_device_participant
+
+            # Assign device to sessions
+            device_session = TeraSession()
+            device_session.id_creator_device = id_device
+            device_session.id_session_type = 1
+            device_session.session_name = 'Creator device session'
+            TeraSession.insert(device_session)
+            id_session = device_session.id_session
+
+            device_session = TeraSession()
+            device_session.id_creator_service = 1
+            device_session.id_session_type = 1
+            device_session.session_name = "Device invitee session"
+            device_session.session_devices = [device]
+            TeraSession.insert(device_session)
+            id_session_invitee = device_session.id_session
+
+            # Attach asset
+            asset = TeraAsset()
+            asset.asset_name = "Device asset test"
+            asset.id_device = id_device
+            asset.id_session = id_session
+            asset.asset_service_uuid = TeraService.get_openteraserver_service().service_uuid
+            asset.asset_type = 'Test'
+            TeraAsset.insert(asset)
+            id_asset = asset.id_asset
+
+            # ... and test
+            test = TeraTest()
+            test.id_device = id_device
+            test.id_session = id_session
+            test.id_test_type = 1
+            test.test_name = "Device test test!"
+            TeraTest.insert(test)
+            id_test = test.id_test
+
+            # Create service config for device
+            device_service_config = TeraServiceConfig()
+            device_service_config.id_device = id_device
+            device_service_config.id_service = 2
[…]
