### 2023-07-13 912fc54
Refs #203. Implemented hard delete for participant
