+++ teraserver/python/tests/opentera/db/models/test_TeraParticipant.py
@@ -1,5 +1,10 @@
 from opentera.db.models.TeraParticipant import TeraParticipant
 from opentera.db.models.TeraParticipantGroup import TeraParticipantGroup
+from opentera.db.models.TeraSession import TeraSession
+from opentera.db.models.TeraAsset import TeraAsset
+from opentera.db.models.TeraTest import TeraTest
+from opentera.db.models.TeraService import TeraService
+
 import uuid
 from tests.opentera.db.models.BaseModelsTest import BaseModelsTest
 
@@ -34,12 +39,96 @@ class TeraParticipantTest(BaseModelsTest):
             loadedParticipant = TeraParticipant.get_participant_by_token(token)
             self.assertEqual(loadedParticipant.participant_uuid, participant.participant_uuid)
 
-    def test_json(self):
+    def test_soft_delete(self):
         with self._flask_app.app_context():
-            return
-            participant = TeraParticipant.get_participant_by_name('Participant #1')
+            # Create a new participant
+            participant = TeraParticipant()
+            participant.participant_name = "Test participant"
+            participant.id_project = 1
+            TeraParticipant.insert(participant)
+            self.assertIsNotNone(participant.id_participant)
+            id_participant = participant.id_participant
 
-            json = participant.to_json()
+            # Delete participant
+            TeraParticipant.delete(id_participant)
+            # Make sure participant is deleted
+            self.assertIsNone(TeraParticipant.get_participant_by_id(id_participant))
 
-            print(json)
+            # Query, with soft delete flag
+            participant = TeraParticipant.query.filter_by(id_participant=id_participant)\
+                .execution_options(include_deleted=True).first()
+            self.assertIsNotNone(participant)
+            self.assertIsNotNone(participant.deleted_at)
+
+    def test_hard_delete(self):
+        with self._flask_app.app_context():
+            # Create a new participant
+            participant = TeraParticipant()
+            participant.participant_name = "Test participant"
+            participant.id_project = 1
+            TeraParticipant.insert(participant)
+            self.assertIsNotNone(participant.id_participant)
+            id_participant = participant.id_participant
+
+            # Assign participant to sessions
+            part_session = TeraSession()
+            part_session.id_creator_participant = id_participant
+            part_session.id_session_type = 1
+            part_session.session_name = 'Creator participant session'
+            TeraSession.insert(part_session)
+            id_session = part_session.id_session
+
+            part_session = TeraSession()
+            part_session.id_creator_service = 1
+            part_session.id_session_type = 1
+            part_session.session_name = "Participant invitee session"
+            part_session.session_participants = [participant]
+            TeraSession.insert(part_session)
+            id_session_invitee = part_session.id_session
+
+            # Attach asset
+            asset = TeraAsset()
+            asset.asset_name = "Participant asset test"
+            asset.id_participant = id_participant
+            asset.id_session = id_session
+            asset.asset_service_uuid = TeraService.get_openteraserver_service().service_uuid
+            asset.asset_type = 'Test'
+            TeraAsset.insert(asset)
+            id_asset = asset.id_asset
+
+            # ... and test
+            test = TeraTest()
+            test.id_participant = id_participant
+            test.id_session = id_session
+            test.id_test_type = 1
+            test.test_name = "Device test test!"
+            TeraTest.insert(test)
+            id_test = test.id_test
+
+            # Soft delete device to prevent relationship integrity errors as we want to test hard-delete cascade here
+            TeraSession.delete(id_session)
+            TeraSession.delete(id_session_invitee)
[…]
