+++ teraserver/python/tests/opentera/db/models/BaseModelsTest.py
@@ -15,6 +15,7 @@ class BaseModelsTest(unittest.TestCase):
         cls._flask_app.config.update({'PROPAGATE_EXCEPTIONS': True})
         cls._db_man = DBManager(cls._config, app=cls._flask_app)
         # Setup DB in RAM
+        # cls._db_man.open_local({'filename': 'D:\\temp\\opentera.db'}, echo=False, ram=False)
         cls._db_man.open_local({}, echo=False, ram=True)
 
         # Creating default users / tests. Time-consuming, only once per test file.
