+++ teraserver/python/tests/opentera/db/models/test_TeraUserGroup.py
@@ -227,11 +227,67 @@ class TeraUserGroupTest(BaseModelsTest):
                     self.assertRaises(SQLAlchemyError, TeraUserGroup.update, group.id_user_group, invalid_fields)
                     TeraUserGroup.db().session.rollback()
 
+    def test_soft_delete(self):
+        with self._flask_app.app_context():
+            # Create new
+            ug = TeraUserGroup()
+            ug.user_group_name = "Test User Group"
+            TeraUserGroup.insert(ug)
+            self.assertIsNotNone(ug.id_user_group)
+            id_user_group = ug.id_user_group
+
+            # Soft delete
+            TeraUserGroup.delete(id_user_group)
+            # Make sure participant is deleted
+            self.assertIsNone(TeraUserGroup.get_user_group_by_id(id_user_group))
+
+            # Query, with soft delete flag
+            ug = TeraUserGroup.query.filter_by(id_user_group=id_user_group).\
+                execution_options(include_deleted=True).first()
+            self.assertIsNotNone(ug)
+            self.assertIsNotNone(ug.deleted_at)
+
     def test_hard_delete(self):
         with self._flask_app.app_context():
-            pass
+            # Create new
+            ug = TeraUserGroup()
+            ug.user_group_name = "Test User Group"
+            TeraUserGroup.insert(ug)
+            self.assertIsNotNone(ug.id_user_group)
+            id_user_group = ug.id_user_group
 
-    # def test_soft_delete(self):
-    #    pass
+            user = TeraUser()
+            user.user_enabled = True
+            user.user_firstname = "Test"
+            user.user_lastname = "User"
+            user.user_profile = ""
+            user.user_password = TeraUser.encrypt_password("test")
+            user.user_superadmin = False
+            user.user_username = "test"
+            user.user_user_groups = [ug]
+            TeraUser.insert(user)
+            self.assertIsNotNone(user.id_user)
+            id_user = user.id_user
+
+            # Soft delete to prevent relationship integrity errors as we want to test hard-delete cascade here
+            id_user_user_group = TeraUserUserGroup.query_user_user_group_for_user_user_group(id_user, id_user_group)\
+                .id_user_user_group
+            TeraUserUserGroup.delete(id_user_user_group)
+            TeraUserGroup.delete(id_user_group)
+
+            # Check that relationships are still there
+            self.assertIsNotNone(TeraUser.get_user_by_id(id_user))
+            self.assertIsNone(TeraUserGroup.get_user_group_by_id(id_user_group))
+            self.assertIsNotNone(TeraUserGroup.get_user_group_by_id(id_user_group, True))
+            self.assertIsNone(TeraUserUserGroup.get_user_user_group_by_id(id_user_user_group))
+            self.assertIsNotNone(TeraUserUserGroup.get_user_user_group_by_id(id_user_user_group, True))
+
+            # Hard delete
+            TeraUserGroup.delete(id_user_group, hard_delete=True)
+
+            # Make sure eveything is deleted
+            self.assertIsNotNone(TeraUser.get_user_by_id(id_user, True))
+            self.assertIsNone(TeraUserGroup.get_user_group_by_id(id_user_group, True))
+            self.assertIsNone(TeraUserUserGroup.get_user_user_group_by_id(id_user_user_group, True))
 
 
