+++ teraserver/python/tests/opentera/db/models/test_TeraService.py
@@ -1,6 +1,7 @@
 from tests.opentera.db.models.BaseModelsTest import BaseModelsTest
 from sqlalchemy import exc
 from opentera.db.models.TeraService import TeraService
+from opentera.db.models.TeraServiceSite import TeraServiceSite
 
 
 class TeraServiceTest(BaseModelsTest):
@@ -330,23 +331,23 @@ class TeraServiceTest(BaseModelsTest):
         #
         """
         return
-        with self._flask_app.app_context():
-            new_service = TeraService()
-            new_service.service_uuid = 'Definitely longer than a 36 characters string'
-            new_service.service_name = 'Name'
-            new_service.service_key = 'key'
-            new_service.service_hostname = 'Hostname'
-            new_service.service_port = 2
-            new_service.service_endpoint = "Endpoint"
-            new_service.service_clientendpoint = 'Clientendpoint'
-            new_service.service_enabled = True
-            new_service.service_system = True
-            new_service.service_editable_config = True
-            self.db.session.add(new_service)
-            self.db.session.commit()
-            self.assertRaises(exc.IntegrityError, self.db.session.commit)
+        # with self._flask_app.app_context():
+        #     new_service = TeraService()
+        #     new_service.service_uuid = 'Definitely longer than a 36 characters string'
+        #     new_service.service_name = 'Name'
+        #     new_service.service_key = 'key'
+        #     new_service.service_hostname = 'Hostname'
+        #     new_service.service_port = 2
+        #     new_service.service_endpoint = "Endpoint"
+        #     new_service.service_clientendpoint = 'Clientendpoint'
+        #     new_service.service_enabled = True
+        #     new_service.service_system = True
+        #     new_service.service_editable_config = True
+        #     self.db.session.add(new_service)
+        #     self.db.session.commit()
+        #     self.assertRaises(exc.IntegrityError, self.db.session.commit)
 
-    def test_service_port_integer(self):
+    def test_service_port_integer_value(self):
         """
         #
         # SQLite uses what it calls a dynamic typing system, which ultimately means that you can store text
@@ -354,20 +355,84 @@ class TeraServiceTest(BaseModelsTest):
         # attempts to do this will fail - not with SQLite.
         """
         return
+        # with self._flask_app.app_context():
+        #     new_service = TeraService()
+        #
+        #     new_service.service_uuid = 'uuid'
+        #     new_service.service_name = 'Name'
+        #     new_service.service_key = 'key'
+        #     new_service.service_hostname = 'Hostname'
+        #
+        #     new_service.service_port = 'not an integer'
+        #
+        #     new_service.service_endpoint = "Endpoint"
+        #     new_service.service_clientendpoint = 'Clientendpoint'
+        #     new_service.service_enabled = True
+        #     new_service.service_system = True
+        #     new_service.service_editable_config = True
+        #     self.db.session.add(new_service)
+        #     self.db.session.commit()
+
+    def test_soft_delete(self):
         with self._flask_app.app_context():
-            new_service = TeraService()
+            # Create new
+            service = TeraService()
+            service.service_name = 'Test Service'
+            service.service_key = 'TestService'
+            service.service_hostname = 'localhost'
+            service.service_port = 12345
+            service.service_endpoint = 'test'
+            service.service_clientendpoint = '/'
+            TeraService.insert(service)
+            self.assertIsNotNone(service.id_service)
+            id_service = service.id_service
 
-            new_service.service_uuid = 'uuid'
-            new_service.service_name = 'Name'
-            new_service.service_key = 'key'
-            new_service.service_hostname = 'Hostname'
+            # Soft delete
+            TeraService.delete(id_service)
 
-            new_service.service_port = 'not an integer'
+            # Make sure it is deleted
[…]
