### 2023-07-17 65f634f
Fixed issue querying session types as super admin but with session types not associated to projects.
