### 2023-07-21 #5497 @galessiorob
Thanks for clarifying above @doumdi, and understood about the different scopes and applications. I think even just adding a line that acknowledges the previous publication and, as @Rocsg mentions, explains the "diff", just like you did in your comment above, would be very beneficial as there might be users that look up this paper or the package by name and could be confused by the previous one surfacing. I don't see any conflict in having both publications, but I do think it's worth mentioning them.

@NoujoudNader mind updating us with the ETA for your review - thanks!
