### 2023-07-22 #5497 @doumdi
@galessiorob from what I understand, you are suggesting to add more information in the JOSS paper clarifying the difference between the two papers.  I would change this sentence and clarify as mentioned above:  "OpenTera has been deployed for robot teleoperation during COVID [@panchea_opentera_2022] and is currently used for multiple rehabilitation projects."  Should I make the change in another branch of the project in the joss-paper directory ? I am a little confused about the process here and what I need to do. Maybe this justification needs to be written somewhere else?
