### 2023-07-26 cbe916e
Refs #202. Updated undelete behavior
