+++ teraserver/python/tests/opentera/db/models/test_TeraTest.py
@@ -1,5 +1,10 @@
 from tests.opentera.db.models.BaseModelsTest import BaseModelsTest
 from opentera.db.models.TeraTest import TeraTest
+from opentera.db.models.TeraSession import TeraSession
+from opentera.db.models.TeraParticipant import TeraParticipant
+from opentera.db.models.TeraUser import TeraUser
+from opentera.db.models.TeraDevice import TeraDevice
+from sqlalchemy.exc import IntegrityError
 
 
 class TeraTestTest(BaseModelsTest):
@@ -11,12 +16,7 @@ class TeraTestTest(BaseModelsTest):
     def test_soft_delete(self):
         with self._flask_app.app_context():
             # Create new
-            test = TeraTest()
-            test.id_participant = 1
-            test.id_session = 1
-            test.id_test_type = 1
-            test.test_name = "Test test!"
-            TeraTest.insert(test)
+            test = TeraTestTest.new_test_test(id_session=1, id_participant=1)
             id_test = test.id_test
 
             # Soft delete
@@ -33,12 +33,7 @@ class TeraTestTest(BaseModelsTest):
     def test_hard_delete(self):
         with self._flask_app.app_context():
             # Create new
-            test = TeraTest()
-            test.id_participant = 1
-            test.id_session = 1
-            test.id_test_type = 1
-            test.test_name = "Test test!"
-            TeraTest.insert(test)
+            test = TeraTestTest.new_test_test(id_session=1, id_device=1)
             self.assertIsNotNone(test.id_test)
             id_test = test.id_test
 
@@ -50,19 +45,107 @@ class TeraTestTest(BaseModelsTest):
 
     def test_undelete(self):
         with self._flask_app.app_context():
-            # Create new
-            test = TeraTest()
-            test.id_participant = 1
-            test.id_session = 1
-            test.id_test_type = 1
-            test.test_name = "Test test!"
-            TeraTest.insert(test)
+            # Create new participant
+            from test_TeraParticipant import TeraParticipantTest
+            participant = TeraParticipantTest.new_test_participant(id_project=1)
+            id_participant = participant.id_participant
+
+            # Create new device
+            from test_TeraDevice import TeraDeviceTest
+            device = TeraDeviceTest.new_test_device()
+            id_device = device.id_device
+
+            # Create new user
+            from test_TeraUser import TeraUserTest
+            user = TeraUserTest.new_test_user()
+            id_user = user.id_user
+
+            # Create new session
+            from test_TeraSession import TeraSessionTest
+            ses = TeraSessionTest.new_test_session(participants=[participant], users=[user], devices=[device])
+            id_session = ses.id_session
+
+            # Create new test
+            test = TeraTestTest.new_test_test(id_session=ses.id_session, id_user=id_user, id_participant=id_participant,
+                                              id_device=id_device)
             self.assertIsNotNone(test.id_test)
             id_test = test.id_test
 
-            # Soft delete
-            TeraTest.delete(id_test)
+            # Undelete
+            TeraTest.undelete(id_test)
 
-            # Make sure it is deleted
-            self.assertIsNone(TeraTest.get_test_by_id(id_test))
+            # Make sure it is back!
+            self.db.session.expire_all()
+            test = TeraTest.get_test_by_id(id_test)
+            self.assertIsNotNone(test)
+            self.assertIsNone(test.deleted_at)
 
+            # Now, delete again but with its dependencies...
+            # Test will be deleted with the session
+            TeraSession.delete(id_session)
+            TeraParticipant.delete(id_participant)
+            TeraDevice.delete(id_device)
+            TeraUser.delete(id_user)
+
+            # Exception should be thrown when trying to undelete
+            with self.assertRaises(IntegrityError) as cm:
[…]
