+++ teraserver/python/tests/opentera/db/models/test_TeraDevice.py
@@ -8,6 +8,7 @@ from opentera.db.models.TeraDeviceSite import TeraDeviceSite
 from opentera.db.models.TeraDeviceParticipant import TeraDeviceParticipant
 from opentera.db.models.TeraServiceConfig import TeraServiceConfig
 from opentera.db.models.TeraSession import TeraSession
+
 from tests.opentera.db.models.BaseModelsTest import BaseModelsTest
 
 
@@ -114,40 +115,27 @@ class TeraDeviceTest(BaseModelsTest):
             id_device = device.id_device
 
             # Assign device to site
-            device_site = TeraDeviceSite()
-            device_site.id_site = 1
-            device_site.id_device = id_device
-            TeraDeviceSite.insert(device_site)
+            from test_TeraDeviceSite import TeraDeviceSiteTest
+            device_site = TeraDeviceSiteTest.new_test_device_site(id_device=id_device, id_site=1)
             id_device_site = device_site.id_device_site
 
             # Assign device to project
-            device_project = TeraDeviceProject()
-            device_project.id_device = id_device
-            device_project.id_project = 1
-            TeraDeviceProject.insert(device_project)
+            from test_TeraDeviceProject import TeraDeviceProjectTest
+            device_project = TeraDeviceProjectTest.new_test_device_project(id_device=id_device, id_project=1)
             id_device_project = device_project.id_device_project
 
             # Assign device to participants
-            device_participant = TeraDeviceParticipant()
-            device_participant.id_device = id_device
-            device_participant.id_participant = 1
-            TeraDeviceParticipant.insert(device_participant)
+            from test_TeraDeviceParticipant import TeraDeviceParticipantTest
+            device_participant = TeraDeviceParticipantTest.new_test_device_participant(id_device=id_device,
+                                                                                       id_participant=1)
             id_device_participant = device_participant.id_device_participant
 
             # Assign device to sessions
-            device_session = TeraSession()
-            device_session.id_creator_device = id_device
-            device_session.id_session_type = 1
-            device_session.session_name = 'Creator device session'
-            TeraSession.insert(device_session)
+            from test_TeraSession import TeraSessionTest
+            device_session = TeraSessionTest.new_test_session(id_creator_device=id_device)
             id_session = device_session.id_session
 
-            device_session = TeraSession()
-            device_session.id_creator_service = 1
-            device_session.id_session_type = 1
-            device_session.session_name = "Device invitee session"
-            device_session.session_devices = [device]
-            TeraSession.insert(device_session)
+            device_session = TeraSessionTest.new_test_session(id_creator_service=1, devices=[device])
             id_session_invitee = device_session.id_session
 
             # Attach asset
@@ -158,19 +146,13 @@ class TeraDeviceTest(BaseModelsTest):
             id_asset = asset.id_asset
 
             # ... and test
-            test = TeraTest()
-            test.id_device = id_device
-            test.id_session = id_session
-            test.id_test_type = 1
-            test.test_name = "Device test test!"
-            TeraTest.insert(test)
+            from test_TeraTest import TeraTestTest
+            test = TeraTestTest.new_test_test(id_session=id_session, id_device=id_device)
             id_test = test.id_test
 
             # Create service config for device
-            device_service_config = TeraServiceConfig()
-            device_service_config.id_device = id_device
-            device_service_config.id_service = 2
-            TeraServiceConfig.insert(device_service_config)
+            from test_TeraServiceConfig import TeraServiceConfigTest
+            device_service_config = TeraServiceConfigTest.new_test_service_config(id_device=id_device, id_service=2)
             id_service_config = device_service_config.id_service_config
 
[…]
