+++ teraserver/python/tests/opentera/db/models/test_TeraParticipant.py
@@ -4,6 +4,9 @@ from opentera.db.models.TeraSession import TeraSession
 from opentera.db.models.TeraAsset import TeraAsset
 from opentera.db.models.TeraTest import TeraTest
 from opentera.db.models.TeraService import TeraService
+from opentera.db.models.TeraProject import TeraProject
+from opentera.db.models.TeraServiceConfig import TeraServiceConfig
+from sqlalchemy.exc import IntegrityError
 
 import uuid
 from tests.opentera.db.models.BaseModelsTest import BaseModelsTest
@@ -65,38 +68,23 @@ class TeraParticipantTest(BaseModelsTest):
             id_participant = participant.id_participant
 
             # Assign participant to sessions
-            part_session = TeraSession()
-            part_session.id_creator_participant = id_participant
-            part_session.id_session_type = 1
-            part_session.session_name = 'Creator participant session'
-            TeraSession.insert(part_session)
+            from test_TeraSession import TeraSessionTest
+            part_session = TeraSessionTest.new_test_session(id_creator_participant=id_participant)
             id_session = part_session.id_session
 
-            part_session = TeraSession()
-            part_session.id_creator_service = 1
-            part_session.id_session_type = 1
-            part_session.session_name = "Participant invitee session"
-            part_session.session_participants = [participant]
-            TeraSession.insert(part_session)
+            part_session = TeraSessionTest.new_test_session(id_creator_service=1, participants=[participant])
             id_session_invitee = part_session.id_session
 
             # Attach asset
-            asset = TeraAsset()
-            asset.asset_name = "Participant asset test"
-            asset.id_participant = id_participant
-            asset.id_session = id_session
-            asset.asset_service_uuid = TeraService.get_openteraserver_service().service_uuid
-            asset.asset_type = 'Test'
-            TeraAsset.insert(asset)
+            from test_TeraAsset import TeraAssetTest
+            asset = TeraAssetTest.new_test_asset(id_session=id_session,
+                                                 service_uuid=TeraService.get_openteraserver_service().service_uuid,
+                                                 id_participant=id_participant)
             id_asset = asset.id_asset
 
             # ... and test
-            test = TeraTest()
-            test.id_participant = id_participant
-            test.id_session = id_session
-            test.id_test_type = 1
-            test.test_name = "Device test test!"
-            TeraTest.insert(test)
+            from test_TeraTest import TeraTestTest
+            test = TeraTestTest.new_test_test(id_session=id_session, id_participant=id_participant)
             id_test = test.id_test
 
             # Soft delete device to prevent relationship integrity errors as we want to test hard-delete cascade here
@@ -126,11 +114,108 @@ class TeraParticipantTest(BaseModelsTest):
             self.assertIsNone(TeraAsset.get_asset_by_id(id_asset, True))
             self.assertIsNone(TeraTest.get_test_by_id(id_test, True))
 
+    def test_undelete(self):
+        with self._flask_app.app_context():
+            # Create a new project
+            from test_TeraProject import TeraProjectTest
+            project = TeraProjectTest.new_test_project()
+            id_project = project.id_project
+
+            # Create a new participant
+            participant = TeraParticipantTest.new_test_participant(id_project=id_project)
+            self.assertIsNotNone(participant.id_participant)
+            id_participant = participant.id_participant
+
+            # Assign participant to sessions
+            from test_TeraSession import TeraSessionTest
+            part_session = TeraSessionTest.new_test_session(id_creator_participant=id_participant)
+            id_session = part_session.id_session
+
+            part_session = TeraSessionTest.new_test_session(id_creator_service=1, participants=[participant])
+            id_session_invitee = part_session.id_session
+
+            # Attach asset
[…]
