+++ teraserver/python/tests/opentera/db/models/test_TeraProject.py
@@ -1,9 +1,11 @@
 from tests.opentera.db.models.BaseModelsTest import BaseModelsTest
 from sqlalchemy import exc
 from opentera.db.models.TeraProject import TeraProject
+from opentera.db.models.TeraSite import TeraSite
 from opentera.db.models.TeraParticipant import TeraParticipant
 from opentera.db.models.TeraDevice import TeraDevice
 from opentera.db.models.TeraParticipantGroup import TeraParticipantGroup
+from sqlalchemy.exc import IntegrityError
 
 
 class TeraProjectTest(BaseModelsTest):
@@ -27,11 +29,7 @@ class TeraProjectTest(BaseModelsTest):
 
     def test_to_json(self):
         with self._flask_app.app_context():
-            new_project = TeraProject()
-            new_project.id_site = 1
-            new_project.project_name = 'test_to_json'
-            self.db.session.add(new_project)
-            self.db.session.commit()
+            new_project = TeraProjectTest.new_test_project()
             new_project_json = new_project.to_json()
             new_project_json_minimal = new_project.to_json(minimal=True)
             self._check_json(new_project, project_test=new_project_json)
@@ -46,71 +44,43 @@ class TeraProjectTest(BaseModelsTest):
 
     def test_to_json_create_event(self):
         with self._flask_app.app_context():
-            new_project = TeraProject()
-            new_project.id_site = 1
-            new_project.project_name = 'test_to_json_create_event'
-            self.db.session.add(new_project)
-            self.db.session.commit()
+            new_project = TeraProjectTest.new_test_project()
             new_project_json = new_project.to_json_create_event()
             self._check_json(new_project, project_test=new_project_json, minimal=True)
 
     def test_to_json_update_event(self):
         with self._flask_app.app_context():
-            new_project = TeraProject()
-            new_project.id_site = 1
-            new_project.project_name = 'test_to_json_update_event'
-            self.db.session.add(new_project)
-            self.db.session.commit()
+            new_project = TeraProjectTest.new_test_project()
             new_project_json = new_project.to_json_update_event()
             self._check_json(new_project, project_test=new_project_json, minimal=True)
 
     def test_to_json_delete_event(self):
         with self._flask_app.app_context():
-            new_project = TeraProject()
-            new_project.id_site = 1
-            new_project.project_name = 'test_to_json_delete_event'
-            self.db.session.add(new_project)
-            self.db.session.commit()
+            new_project = TeraProjectTest.new_test_project()
             new_project_json = new_project.to_json_delete_event()
             self.assertGreaterEqual(new_project_json['id_project'], 1)
 
     def test_get_users_ids_in_project(self):
         with self._flask_app.app_context():
-            new_project = TeraProject()
-            new_project.id_site = 1
-            new_project.project_name = 'test_get_users_ids_in_project'
-            self.db.session.add(new_project)
-            self.db.session.commit()
+            new_project = TeraProjectTest.new_test_project()
             users_ids = new_project.get_users_ids_in_project()
             self.assertIsNotNone(users_ids)
 
     def test_get_users_in_project(self):
         with self._flask_app.app_context():
-            new_project = TeraProject()
-            new_project.id_site = 1
-            new_project.project_name = 'test_get_users_in_project'
-            self.db.session.add(new_project)
-            self.db.session.commit()
+            new_project = TeraProjectTest.new_test_project()
             users = new_project.get_users_in_project()
             self.assertIsNotNone(users)
 
     def test_get_project_by_projectname(self):
         with self._flask_app.app_context():
-            new_project = TeraProject()
-            new_project.id_site = 1
-            new_project.project_name = 'test_get_project_by_projectname'
-            self.db.session.add(new_project)
-            self.db.session.commit()
+            new_project = TeraProjectTest.new_test_project(name='test_project_by_name')
[…]
