+++ teraserver/python/tests/opentera/db/models/test_TeraTestType.py
@@ -1,6 +1,8 @@
 from tests.opentera.db.models.BaseModelsTest import BaseModelsTest
 from opentera.db.models.TeraTestType import TeraTestType
 from opentera.db.models.TeraTest import TeraTest
+from opentera.db.models.TeraTestTypeSite import TeraTestTypeSite
+from opentera.db.models.TeraTestTypeProject import TeraTestTypeProject
 
 
 class TeraTestTypeTest(BaseModelsTest):
@@ -12,10 +14,7 @@ class TeraTestTypeTest(BaseModelsTest):
     def test_soft_delete(self):
         with self._flask_app.app_context():
             # Create new
-            test_type = TeraTestType()
-            test_type.test_type_name = "Test test type..."
-            test_type.id_service = 1
-            TeraTestType.insert(test_type)
+            test_type = TeraTestTypeTest.new_test_test_type()
             id_test_type = test_type.id_test_type
 
             # Soft delete
@@ -33,18 +32,11 @@ class TeraTestTypeTest(BaseModelsTest):
     def test_hard_delete(self):
         with self._flask_app.app_context():
             # Create new
-            test_type = TeraTestType()
-            test_type.test_type_name = "Test test type..."
-            test_type.id_service = 1
-            TeraTestType.insert(test_type)
+            test_type = TeraTestTypeTest.new_test_test_type()
             id_test_type = test_type.id_test_type
 
-            test = TeraTest()
-            test.id_participant = 1
-            test.id_session = 1
-            test.id_test_type = id_test_type
-            test.test_name = "Test test!"
-            TeraTest.insert(test)
+            from test_TeraTest import TeraTestTest
+            test = TeraTestTest.new_test_test(id_session=1, id_participant=1, id_test_type=id_test_type)
             self.assertIsNotNone(test.id_test)
             id_test = test.id_test
 
@@ -64,3 +56,50 @@ class TeraTestTypeTest(BaseModelsTest):
             # Make sure eveything is deleted
             self.assertIsNone(TeraTest.get_test_by_id(id_test, True))
             self.assertIsNone(TeraTestType.get_test_type_by_id(id_test_type, True))
+
+    def test_undelete(self):
+        with self._flask_app.app_context():
+            # Create new
+            test_type = TeraTestTypeTest.new_test_test_type()
+            id_test_type = test_type.id_test_type
+
+            from test_TeraTest import TeraTestTest
+            test = TeraTestTest.new_test_test(id_session=1, id_participant=1, id_test_type=id_test_type)
+            self.assertIsNotNone(test.id_test)
+            id_test = test.id_test
+
+            # Associate with site
+            from test_TeraTestTypeSite import TeraTestTypeSiteTest
+            tt_site = TeraTestTypeSiteTest.new_test_test_type_site(id_site=1, id_test_type=id_test_type)
+            id_test_type_site = tt_site.id_test_type_site
+
+            # ... and project
+            from test_TeraTestTypeProject import TeraTestTypeProjectTest
+            tt_project = TeraTestTypeProjectTest.new_test_test_type_project(id_project=1, id_test_type=id_test_type)
+            id_test_type_project = tt_project.id_test_type_project
+
+            # Delete
+            TeraTest.delete(id_test)
+            TeraTestType.delete(id_test_type)
+
+            # Check that everything was deleted
+            self.assertIsNone(TeraTest.get_test_by_id(id_test))
+            self.assertIsNone(TeraTestType.get_test_type_by_id(id_test_type))
+            self.assertIsNone(TeraTestTypeSite.get_test_type_site_by_id(id_test_type_site))
[…]
