### 2023-07-31 #5497 @galessiorob
@doumdi yes, I think that the clarification you mention is enough to clarify the difference between the two papers. You can make the change either in a branch or since it's so small and [only for the text](https://github.com/introlab/opentera/blob/main/joss-paper/paper.md), directly on your main since you're the author. Editors and reviewers will make suggested changes in issues in your repo, or if they are specific changes in a branch if you are okay with this type of review.
