### 2023-07-31 #5497 @galessiorob
@doumdi I put together a few edit suggestions for the paper [here](https://github.com/introlab/opentera/issues/225), please adapt as you see fit. I have emailed @NoujoudNader to ask what their ETA on the review is.
