### 2023-08-29 d60d787
Refs #225, review paper and suggestions. Will run GitHub action to generate paper as artifact on branch joss-paper-review.
