+++ joss-paper/paper.md
@@ -67,34 +67,37 @@ bibliography: paper.bib
 
 # Summary
 
-OpenTera is a microservice based framework primarily developed to support telehealth research projects and real-world deployment. This project is based on 20 years of experience linking at-home participants to remote users (such as clinicians, researchers, healthcare professionals) with audio-video-data connections and in-the-field sensors, such as biometrics, wearable and robotics devices. Applications of the OpenTera framework are not limited to research projects and can also be used in clinical environments.
-Most telehealth-based research projects require a common data structure: data collection sites, projects, participants and sessions including various recorded data types from sensors or other sources. They also require many common features: user authentication based on various access roles, ability to add new features based on specific project needs, ease of use for the participant, and secure data hosting.  These features are also shared between research projects: videoconferencing with specific health related features (e.g., angles measurement, timers), surveys data collection, data analysis and exportation.
+OpenTera is a microservice-based framework primarily developed to support telehealth research projects and real-world deployment. This project is based on 20 years of experience linking at-home participants to remote users (such as clinicians, researchers, healthcare and professionals) with audio-video-data connections and in-the-field sensors, such as biometrics, wearable and robotics devices. Applications of the OpenTera framework are not limited to research projects and can also be used in clinical environments.
 
-Many of the available solutions are either costly, feature limited, proprietary (e.g., can hardly be adapted for research purposes and raw data is harder to access) or hard to deploy in a telehealth context. OpenTera was built for extensibility to provide research projects full control over their data and hosting.
+Most telehealth-based research projects require a common data structure: data collection sites, projects, participants and sessions, including various recorded data types from sensors or other sources. They also require standard features: user authentication based on various access roles, the ability to add new features based on specific project needs, ease of use for the participant, and secure data hosting. These features are also shared between research projects: videoconferencing with specific health-related features (e.g., angles measurement, timers), surveys data collection, data analysis and exportation.
+
+Many available solutions are costly, feature limited, proprietary (e.g., can hardly be adapted for research purposes and raw data is harder to access) or hard to deploy in a telehealth context. OpenTera was built for extensibility to provide research projects complete control over their data and hosting.
 
 # Statement of need
 
 From our research experience, common features between the different telehealth projects emerged:
 
 * **Data structure.** Store data in a structured way to ease data extraction and analysis.
-* **Ecological data capture.** Collect data not only in laboratories or controlled environments, but also in homes or institutions.
-* **Project adaptability.** Develop project-specific dashboards and user interfaces while reusing what was previously implemented as much as possible of to reduce development time. Rehabilitation projects may require implementing serious games or exergames, while teleoperation projects may require real-time navigation tools. Adapting already existing open-source software when possible is often the key.
-* **Cost effectiveness.** Most of the recent commercial cloud telehealth applications available are subscription-based and do not offer the flexibility needed. Each vendor offers its own approach tailored for its products and services. We often have data collections from dozens of participants and users, and paying subscription fees would be prohibitive.
+* **Ecological data capture.** Collect data not only in laboratories or controlled environments but also in homes or institutions.
+* **Project adaptability.** Develop project-specific dashboards and user interfaces while reusing what was previously implemented as much as possible to reduce development time. Rehabilitation projects may require implementing serious games or exergames, while teleoperation projects may require real-time navigation tools. Adapting already existing open-source software when possible is often the key.
[…]
