+++ teraserver/python/env/requirements.txt
@@ -1,41 +1,41 @@
 pypiwin32==223; sys_platform == 'win32'
-Twisted==22.10.0
+Twisted==23.8.0
 treq==22.2.0
-cryptography==41.0.2
+cryptography==41.0.3
 autobahn==23.6.2
-SQLAlchemy==2.0.18
+SQLAlchemy==2.0.20
 sqlalchemy-schemadisplay==1.3
 pydot==1.4.2
-psycopg2-binary==2.9.6
-Flask==2.3.2
-Flask-SQLAlchemy==3.0.5
+psycopg2-binary==2.9.7
+Flask==2.3.3
+Flask-SQLAlchemy==3.1.1
 Flask-Login==0.6.2
 Flask-Login-Multi==0.1.2
 Flask-HTTPAuth==4.8.0
-Flask-SocketIO==5.3.4
+Flask-SocketIO==5.3.6
 Flask-Session==0.5.0
 flask-restx==1.1.0
 Flask-Security==3.0.0
 Flask-Babel==3.1.0
 Flask-BabelEx==0.9.4
-Flask-Migrate==4.0.4
+Flask-Migrate==4.0.5
 flask-swagger-ui==4.11.1
-Flask-Limiter==3.3.1
+Flask-Limiter==3.5.0
 Flask-Mail==0.9.1
 Flask-Principal==0.4.0
-redis==4.6.0
+redis==5.0.0
 txredisapi==1.4.10
 passlib==1.7.4
 bcrypt==4.0.1
 WTForms==3.0.1
 pyOpenSSL==23.2.0
 service-identity==23.1.0
-PyJWT==2.7.0
+PyJWT==2.8.0
 pylzma==0.5.0
 bz2file==0.98
 python-slugify==8.0.1
-websocket-client==1.6.1
-pytest==7.4.0
-jsonschema==4.18.0
+websocket-client==1.6.3
+pytest==7.4.2
+jsonschema==4.19.0
 Jinja2==3.1.2
 ua-parser==0.18.0
