### 2023-09-13 403f613
Refs #227, #228 Implementing any or all user roles, simplified token structure, to be fully tested
