+++ teraserver/python/opentera/services/ServiceAccessManager.py
@@ -428,45 +428,43 @@ class ServiceAccessManager:
             return decorated
         return wrap
 
+    @staticmethod
+    def service_user_roles_any_required(roles: List[str]):
+        def wrap(f):
+            @wraps(f)
+            def decorated(*args, **kwargs):
 
-@staticmethod
-def service_user_roles_any_required(roles: List[str]):
-    def wrap(f):
-        @wraps(f)
-        def decorated(*args, **kwargs):
+                # Check if service is initialized
+                if ServiceAccessManager.service is None or 'service_key' \
+                        not in ServiceAccessManager.service.service_info:
+                    return gettext('Forbidden'), 403
 
-            # Check if service is initialized
-            if ServiceAccessManager.service is None or 'service_key' \
-                    not in ServiceAccessManager.service.service_info:
-                return gettext('Forbidden'), 403
+                service_key = ServiceAccessManager.service.service_info['service_key']
 
-            service_key = ServiceAccessManager.service.service_info['service_key']
+                # Check if user is logged in, watch out not None object but LocalProxy cannot use is None...
+                if not current_user_client:
+                    return gettext('Forbidden'), 403
 
-            # Check if user is logged in, watch out not None object but LocalProxy cannot use is None...
-            if not current_user_client:
-                return gettext('Forbidden'), 403
+                # Super admin pass through
+                if current_user_client.user_superadmin:
+                    return f(*args, **kwargs)
 
-            # Super admin pass through
-            if current_user_client.user_superadmin:
+                # Check if user has the required role (global roles are stored in token)
+                user_roles_from_token = current_user_client.get_roles_for_service(service_key)
+
+                # Check if user has the required roles
+                if not any(role in user_roles_from_token for role in roles):
+                    return gettext('Forbidden'), 403
+
+                # Everything ok, continue
                 return f(*args, **kwargs)
 
-            # Check if user has the required role (global roles are stored in token)
-            user_roles_from_token = current_user_client.get_roles_for_service(service_key)
+            return decorated
 
-            # Check if user has the required roles
-            if not any(role in user_roles_from_token for role in roles):
-                return gettext('Forbidden'), 403
+        return wrap
 
-            # Everything ok, continue
-            return f(*args, **kwargs)
-
-        return decorated
-
-    return wrap
-
-
-@staticmethod
-def service_user_roles_required(roles: List[str]):
-    # For compatibility with old code
-    return ServiceAccessManager.service_user_roles_all_required(roles)
+    @staticmethod
+    def service_user_roles_required(roles: List[str]):
+        # For compatibility with old code
+        return ServiceAccessManager.service_user_roles_all_required(roles)
 
