### 2023-09-13 d27db9d
Refs #227, #228 Implementing any or all user roles, simplified token structure, to be fully tested
