### 2023-09-14 ff691b7
Refs #230. Added first implementation music streaming in VideoRehab.
