+++ teraserver/easyrtc/static/js/tera_webrtc.js
@@ -374,7 +374,7 @@ function localVideoStreamSuccess(stream){
                 for (let i=0; i<remoteStreams.length; i++){
                     easyrtc.addStreamToCall(remoteStreams[i].peerid, stream.streamName, function (/*caller, streamName*/) {
                         //console.log("Added stream to " + caller + " - " + streamName);
-                        updateUserLocalViewLayout(localStreams.length, remoteStreams.length);
+                        updateUserLocalViewLayout(getVideoStreamsCount(localStreams), getVideoStreamsCount(remoteStreams));
                     });
                 }
             }
@@ -552,6 +552,10 @@ function newStreamStarted(callerid, stream, streamname) {
             // Screen sharing = no controls
             showStatusControls(false, slot, false);
         }
+        if (streamname.endsWith("ScreenShareAudio")){
+            // Only sharing audio, video track is always enabled - disable!
+            stream.getVideoTracks()[0].enabled = false;
+        }
     }else{
         showStatusControls(false, slot, true);
         playSound("audioConnected");
@@ -565,11 +569,12 @@ function newStreamStarted(callerid, stream, streamname) {
         //}
     }
 
-    easyrtc.setVideoObjectSrc(getVideoWidget(false, slot)[0], stream);
+    if (stream.getVideoTracks()[0].enabled)
+        easyrtc.setVideoObjectSrc(getVideoWidget(false, slot)[0], stream);
 
     // Update display
-    updateUserRemoteViewsLayout(remoteStreams.length);
-    updateUserLocalViewLayout(localStreams.length, remoteStreams.length);
+    updateUserRemoteViewsLayout(getVideoStreamsCount(remoteStreams));
+    updateUserLocalViewLayout(getVideoStreamsCount(localStreams), getVideoStreamsCount(remoteStreams));
     refreshRemoteStatusIcons(callerid);
 
     if (streamname === "default"){
@@ -722,8 +727,8 @@ function streamDisconnected(callerid, mediaStream, streamName){
         }
     }
 
-    updateUserRemoteViewsLayout(remoteStreams.length);
-    updateUserLocalViewLayout(localStreams.length, remoteStreams.length);
+    updateUserRemoteViewsLayout(getVideoStreamsCount(remoteStreams));
+    updateUserLocalViewLayout(getVideoStreamsCount(localStreams), getVideoStreamsCount(remoteStreams));
 }
 
 function disconnectedFromSignalingServer(){
@@ -1149,20 +1154,26 @@ function signalingLoginFailure(errorCode, message) {
     clearStatusMsg();
 }
 
-async function shareScreen(local, start){
+async function shareScreen(local, start, sound_only = false){
     let streamName = localContact.peerid + '_' +'ScreenShare';
+    if (sound_only){
+        streamName += "Audio";
+    }
     if (start === true){
         // Start screen sharing
         let screenStream = undefined;
         try {
-            // TODO: Check if video can put to false or set "enabled=false"
             screenStream = await navigator.mediaDevices.getDisplayMedia({video: true,
-                audio: currentConfig.screenAudio});
+                audio: sound_only || currentConfig.screenAudio});
+            if (sound_only){
+                // Video track must be stopped if we want sound only, since it is required to get Display Media to share
+                screenStream.getVideoTracks()[0].enabled = false;
+            }
             easyrtc.register3rdPartyLocalMediaStream(screenStream, streamName);
             // TODO: Use easyrtc.setSdpFilters to improve audio quality
-            // easyrtc.setSdpFilters()
+            //easyrtc.setSdpFilters()
 
-            console.log("Starting screen sharing - with audio: " + currentConfig.screenAudio);
[…]
