+++ teraserver/python/config/TeraServerConfig.ini
@@ -10,7 +10,7 @@
 		"ca_certificate": "ca_cert.pem",
 		"ca_private_key": "ca_key.pem",
 		"upload_path": "uploads",
-		"debug_mode": false,
+		"debug_mode": true,
 		"enable_docs": true
 	},	
 	"Database": {
