### 2023-09-14 1110eb6
Refs #230. Fixed audio sharing not working and not stopping. Added constraints to try to improve audio quality
