### 2023-09-14 eadbddb
Refs #230. Fixed screen sharing not showing screen if not audio only.
