### 2023-09-14 44d130c
Refs #230. Fix for not being able to stop music sharing.
