### 2023-09-15 d178fe5
Refs #230. Fixed screen sharing not working anymore. Updated music playing icon.
