### 2023-09-26 fb4d92c
Refs #230. Adjusted video parameters when streaming music.
