### 2023-09-26 2d12479
Refs #230. Tentative fix to mute speakers when remote streams count > 4.
