### 2023-10-04 b33ea06
Refs #230. Reworked layout manager to prevent display of "blank" music video
