+++ teraserver/easyrtc/static/js/tera_ui.js
@@ -335,8 +335,8 @@ function enlargeView(local, index){
     }
 
     // Update layouts
-    updateUserLocalViewLayout(getVideoStreamsCount(localStreams), getVideoStreamsCount(remoteStreams));
-    updateUserRemoteViewsLayout(getVideoStreamsCount(remoteStreams));
+    updateUserLocalViewLayout();
+    updateUserRemoteViewsLayout();
 
 }
 
