### 2023-10-04 4dbaaa7
Refs #230. Tentative fix for unmute speaker not working on music share.
