### 2023-10-05 9141d4d
Refs #230. Added icon and moved video area on participant layout when user is sharing music.
