+++ teraserver/easyrtc/static/js/tera_layout_participants.js
@@ -65,8 +65,13 @@ function updateUserRemoteViewsLayout(){
 
     // Hide unused views
     for (let i=1; i<=maxRemoteSourceNum; i++){
-        if (!usedRemoteVideosIndexes.includes(i))
+        if (!usedRemoteVideosIndexes.includes(i)){
+            if (i<remoteStreams.length){
+                if (remoteStreams[i-1].streamname.endsWith("ShareAudio"))
+                    continue; // Audio stream videos must still be displayed, otherwise mute won't properly work.
+            }
             $("#remoteView" + i).hide();
+        }
     }
 
     // Show used views
@@ -156,6 +161,29 @@ function setLargeView(view_id, updateui=true){
     if (updateui){
         updateUserRemoteViewsLayout();
         updateUserLocalViewLayout();
+    }
+}
+
+function setRemoteStatusVideo(stream_index, set){
+    let video = getVideoWidget(false, stream_index+1);
+    let parent_video = $("#remoteStatusVideo" + (getStreamIndexForPeerId(remoteStreams[stream_index].peerid)+1));
+    if (set){
+        console.log("Setting status video for " + remoteStreams[stream_index].peerid);
+        // Ensure that this video stream is within the correct picture-in-picture
+        if (parent_video[0] && video[0].parentElement.parentElement.id !== parent_video[0].parentElement.id){
+            // Move the video to the correct place
+            parent_video.append(video);
+            showElement(parent_video[0].parentElement.id);
+        }
+    }else{
+        if (parent_video[0]){
+            if (video[0].parentElement.parentElement.id === parent_video[0].parentElement.id){
+                console.log("Removing status video for " + remoteStreams[stream_index].peerid);
+                let prev_el = $('#remoteView' + (stream_index+1));
+                prev_el.append(video);
+                hideElement(parent_video[0].parentElement.id);
+            }
+        }
     }
 }
 
