### 2023-10-12 fc7fe31
Refs #230. Fixed issue on music not playing when speaker is initially muted on Chrome.
