### 2023-10-24 #5497 @Kevin-Mattheus-Moerman
@doumdi are you able to work on the above :point_up: comments/questions from the reviewers/editor? To avoid general delays, and to avoid loosing track of the reviewers, we recommend that you respond to reviewer comments/issues in a timely manor. If you are no longer interested in pursuing a JOSS publication please let us know as well.
