### 2023-10-24 #5497 @doumdi
@Kevin-Mattheus-Moerman @galessiorob, Yes I have integrated the suggestions from @Rocsg and @galessiorob in this [branch](https://github.com/introlab/opentera/tree/joss-paper-review) as I already mentioned in my comment August 29th. @galessiorob reacted to this post with a heart.  Changes can be tracked [here](https://github.com/introlab/opentera/issues/225). I though the only thing missing was @NoujoudNader's review, that is why I asked for an update three weeks ago.  Did I miss something else?

Best Regards,

Dominic
