### 2023-10-25 #5497 @tmickleydoyle
Hey team 👋 Great work! Two notes:

- The installation process is more complex and requires a decent knowledge of a lot of technologies. It would be great if there were a much simpler way to deploy this application, maybe via Docker? I am putting myself in the shoes of a semi-technical IT person at a medical center, and the number of steps required to install this software would be challenging. There are a lot of opportunities for errors, and unless the user is well versed in digesting error logs, they might get tripped up with the install.
- The writing in the paper might need another pass. For English readers, several sections could be read more smoothly. I am happy to take a pass at editing if you all would like. I will not change the references or the essence of the paper. Let me know! If you do not think this is a necessary step, I can be convinced to check the `Quality of writing` box.

Cheers 🎉
