### 2023-10-25 #5497 @doumdi
@tmickleydoyle Thank you for the quick response and the review. Please make sure you make your review (and pull request) on this branch : https://github.com/introlab/opentera/tree/joss-paper-review. Meanwhile, we will update the documentation with Docker instructions.
Regards,
Dominic
