### 2023-10-25 #5497 @tmickleydoyle
@doumdi, do you mind sending over a version of the paper I can edit? Thank you
