### 2023-10-25 #5497 @doumdi
@tmickleydoyle You can download the files by clicking here: https://github.com/introlab/opentera/tree/joss-paper-review/joss-paper, [paper.md](https://raw.githubusercontent.com/introlab/opentera/dev/joss-paper/paper.md), [paper.bib](https://raw.githubusercontent.com/introlab/opentera/dev/joss-paper/paper.bib). Is this working for you ?
