### 2023-10-25 #5497 @doumdi
Draft of Docker installation can be found here but will be improved and incorporated in the main doc today : https://github.com/introlab/opentera/tree/main/docker/dev
