+++ docs/deployment/DeployWithDockerCompose.md
@@ -1,0 +1,3 @@
+# Deploying with Docker Compose on Windows/Mac/Linux
+
+Bla Bla.
