+++ docs/Deployment.md
@@ -1,3 +1,3 @@
 # Deployment Notes
 
-[Docker Compose](deployment/DeployWithDockerCompose.md) deployment is recommended to start using OpenTera Server for development and tests.
+[Docker Compose](deployment/DeployWithDockerCompose.md) deployment is recommended to start using OpenTera Server for development and tests. [Native installation with Linux](deployment/DeployWithLinuxNative.md) is also possible but requires mor knowledge of Linux systems and related tools.
