+++ docs/deployment/DeployWithDockerCompose.md
@@ -1,131 +1,9 @@
-# Deploying with Docker Compose on Windows/Mac/Linux
+# Deploying with Docker on Linux Servers
 
-[Docker Compose](https://docs.docker.com/compose/) is a tool that simplifies the management of multi-container [Docker](https://docs.docker.com/) applications. It allows you to define and run complex applications using a simple YAML file, specifying the services, networks, and volumes needed, making it easier to orchestrate and manage multiple Docker containers as a single application. An example Docker Compose configuration is provided with OpenTeraServer on GitHub in the [docker/dev](https://github.com/introlab/opentera/tree/main/docker/dev) directory.
+[Docker Compose](https://docs.docker.com/compose/) is a tool that simplifies the management of multi-container [Docker](https://docs.docker.com/) applications. It allows you to define and run complex applications using a simple YAML file, specifying the services, networks, and volumes needed, making it easier to orchestrate and manage multiple Docker containers as a single application.
 
 ## Warnings
 
-The following example is provided for testing on a local machine and is not suited for production.
+If you want a test environment or a dev environment, please have a look at the [Developers Setup for Docker documentation](../developers/Developer-Setup-for-Docker)
 
 ## Requirements
-
-* [Install Docker Desktop for Windows/Linux/Mac](https://www.docker.com/products/docker-desktop/)
-* [Install Visual Studio Code](https://code.visualstudio.com/download)
-* Enough free space on your disk (10Gb+)
-
-### Install Docker Desktop
-
-* Follow installation procedures for your operating system and use default settings. You might need to have Administration rights to proceed with installation.
-
-### Install Visual Studio Code
-
-* Follow installation procedures for your operating system and use default settings. You might need to have Administration rights to proceed with installation.
-
-#### Visual Studio Code Extensions
-
-Launch Visual Studio Code and install the following [extensions](https://code.visualstudio.com/docs/editor/extension-marketplace):
-
-* [Python](https://marketplace.visualstudio.com/items?itemName=ms-python.python)
-* [Dev Containers](https://marketplace.visualstudio.com/items?itemName=ms-vscode-remote.remote-containers)
-* [Docker](https://marketplace.visualstudio.com/items?itemName=ms-azuretools.vscode-docker)
-* [CMake](https://marketplace.visualstudio.com/items?itemName=twxs.cmake)
-* [Markdown](https://marketplace.visualstudio.com/items?itemName=yzhang.markdown-all-in-one)
-
-## Open OpenTera Base Folder in Visual Studio Code
-
-We assume here that you have cloned the [code from GitHub](https://github.com/introlab/opentera.git).
-You must select the root directory of the opentera git project.
-
-### Build the Development Environment with docker-compose
-
-A sample [docker-compose.yml](https://github.com/introlab/opentera/blob/main/docker/dev/docker-compose.yml) file is provided that generates a full environment with the following components running in individual containers :
-
-* Redis server
-* PostgresSQL server
-* NGINX reverse proxy server
-* A certificate generator for https self-signed certificates
-* TeraServer and base services
-
-Procedure:
-
-1. Right-click on the [docker-compose.yml](https://github.com/introlab/opentera/blob/main/docker/dev/docker-compose.yml) and select "Compose Up". This can take several minutes the first time to download all the images and start the containers.
-2. Docker compose will create **volumes** (shared disks) to store your databases, configurations, files as specified in the "volume:" section. Those directories are automatically mounted when the containers are started.
-3. When running for the first time, databases with default values are generated.
-4. **The default user is "admin" with password "admin".**
-5. The opentera source code is copied (not mounted) at the moment. If you make changes in the code, you need to do "Compose Up" again to generate the new container.
-
[…]
