+++ docs/deployment/DeployWithDockerCompose.md
@@ -1,9 +1,591 @@
 # Deploying with Docker on Linux Servers
 
-[Docker Compose](https://docs.docker.com/compose/) is a tool that simplifies the management of multi-container [Docker](https://docs.docker.com/) applications. It allows you to define and run complex applications using a simple YAML file, specifying the services, networks, and volumes needed, making it easier to orchestrate and manage multiple Docker containers as a single application.
+This procedure require Linux server management knowledge and adequate permissions. It is provided as an example and can be changed according to your exact requirements and security constraints. OpenTera's authors cannot take any responsability for the use of the software, as mentioned in the [Apache Licence v2](https://www.apache.org/licenses/LICENSE-2.0.txt).
 
 ## Warnings
 
-If you want a test environment or a dev environment, please have a look at the [Developers Setup for Docker documentation](../developers/Developer-Setup-for-Docker)
+* If you want a test environment or a dev environment, please have a look at the [Developers Setup for Docker documentation](../developers/Developer-Setup-for-Docker).
 
 ## Requirements
+
+* A Linux server installation.
+* [Docker server edition](https://docs.docker.com/engine/install/) for your favorite Linux distribution.
+* Make sure you setup the system with a user member of the "docker" group.
+* Internet connexion
+
+## Setup with docker-compose
+
+[Docker Compose](https://docs.docker.com/compose/) is a tool that simplifies the management of multi-container [Docker](https://docs.docker.com/) applications. It allows you to define and run complex applications using a simple YAML file, specifying the services, networks, and volumes needed, making it easier to orchestrate and manage multiple Docker containers as a single application.
+
+The next sections will give details of a sample docker-compose setup. Edit and change it to fit your needs.
+
+### Installation Steps
+
+#### Step 1: Create Docker Environment Directory
+
+Create a directory with the name of the environment you want to configure and change directory to the newly created directory. Name the directory as you wish. All the next steps will be done in this directory.
+
+```bash
+mkdir opentera-docker-prod
+cd opentera-docker-prod
+```
+
+#### Step 2: Create `docker-compose.yml` File
+
+Create and edit a file named `docker-compose.yml` with the following content:
+
+```docker
+version: '3.8'
+#name: might not work depending on docker version, comment it to use the directory name instead (legacy)
+name: opentera-docker-prod
+services:
+  db:
+    # TODO: Use postgres latest image
+    image: postgres:14.1-alpine
+    restart: always
+    environment:
+      # TODO: Change PostGresSQL default username/password
+      - POSTGRES_USER=postgres
+      - POSTGRES_PASSWORD=postgres
+    volumes:
+      - db:/var/lib/postgresql/data
+      # This script will initialize the DB (once) at first startup.
+      - ./db/init.sql:/docker-entrypoint-initdb.d/create_tables.sql
+    networks:
+    # Name the network as you wish, use the same network name for all containers.
+      - opentera-docker-prod-net
+
+  cache:
+    # TODO Use redis latest image
+    image: redis:6.2-alpine
+    restart: always
+    command: redis-server --save 20 1 --loglevel warning
+    volumes:
[…]
