### 2023-10-27 #5497 @tmickleydoyle
@doumdi, here is my PR for some grammar suggestions.

https://github.com/introlab/opentera/pull/231
