### 2023-10-27 dfd4044
Working prod Docker Example and documentation
