### 2023-10-27 0aaac14
Add verbose flag
