+++ _sources/deployment/DeployWithLinuxNative.md.txt
@@ -1,0 +1,193 @@
+# Deploying OpenTera on a Linux Server
+
+Those setup instructions suppose a pre-installed and configured Ubuntu 20.04 server installation, either in a virtual machine or on a physical machine.
+
+## Pre-requisites
+
+This section configures the depending packages and software before installing the main OpenTera server.
+
+<hr>
+
+### Postgresql
+<hr>
+
+1. Install Postgresql package: `sudo apt-get install postgresql`
+2. Change default `postgres` user password for a more secure installation:
+    ```
+    sudo -u postgres psql
+    ALTER USER postgres PASSWORD 'TypeThePasswordHere';
+    ```
+3. (Optional) Setup a local pgAdmin instance to connect to postgresql database (optionally using a SSH tunnel)
+4. Create `teraagent` user in database (or name of your choice, but will have to be adjusted in the config section below):
+    ```
+    CREATE USER teraagent WITH ENCRYPTED PASSWORD 'TypeUserPasswordHere';
+    ```
+5. Create required database and assign `teraagent` user to them:
+    ```
+    CREATE DATABASE opentera WITH OWNER=teraagent;
+    CREATE DATABASE openterafiles WITH OWNER=teraagent;
+    CREATE DATABASE openteralogs WITH OWNER=teraagent;
+    ```
+6. Don't forget to quit the postgres console: `\q`
+
+<hr>
+
+### Redis
+<hr>
+
+1. Install redis server: `sudo apt-get install redis-server`
+2. (Optional, but strongly recommended) Setup a password to the redis server instance:
+    ```
+    `sudo nano /etc/redis/redis.conf`
+    Edit the line `requirepass` and set your password
+    Save, close and restart the redis server: `sudo systemctl restart redis.service`
+    ```
+
+<hr>
+
+### nginx
+<hr>
+Only basic configuration is done here - specific OpenTera configuration is done below
+
+1. Install nginx: `sudo apt-get install nginx`
+
+<hr>
+
+### Python environment (using miniconda)
+<hr>
+
+1. Follow the instructions [here­](https://docs.conda.io/projects/conda/en/latest/user-guide/install/linux.html) to download and install miniconda
+2. When requested, execute the `conda init` command
+3. Close your shell and restart it again
+
+<hr>
+
+### Build environment
+<hr>
+
+1. Install git: `sudo apt-get install git`
+2. Install cmake: `sudo apt-get install cmake`
+3. Install g++: `sudo apt-get install g++`
+4. Install nodejs / npm: `sudo apt-get install npm`
+
+<hr>
+
+## OpenTera installation
+
+This section proceeds to the installation of the OpenTera server in itself.
+
+<hr>
+
+### Installation
+<hr>
+
+1. Fetch the OpenTera code with submodules using git: `git clone --recurse-submodules https://github.com/introlab/opentera.git`
+2. `cd opentera/teraserver` (or the location that you cloned the project)
+3. Initialize cmake environment: `cmake .`
+4. Generate python environment using make: `make python-all`
+5. Generate nodejs environment for VideoRehab service:
+    ```
+    cd opentera/teraserver/easyrtc
+    npm install
+    ```
+
+<hr>
+
+### Configuration
+<hr>
+
+#### Config files
+There is a few config files to edit. You should edit each of them and put the correct parameters, according to your setup and the passwords you've set previously. Here is the list of the files:
+
+* `teraserver/python/config/TeraServerConfig.ini`: the main config file. "port" and "hostname" shouldn't be changed.
+* `teraserver/python/config/nginx.conf`: nginx config file. Unless listening to a different port and setting correct ssl certificates, nothing should be changed in that file.
[…]
