+++ _sources/deployment/DeployWithDockerCompose.md.txt
@@ -1,0 +1,3 @@
+# Deploying with Docker on Linux Servers
+
+A complete [sample configuration for production is provided on GitHub](https://github.com/introlab/opentera/tree/main/docker/main). You can modify it to fit your needs.
