### 2023-10-27 48d604c
Update TeraServerConfig.ini
Disable debug mode.
