+++ docker/prod/teraserver/TeraServerConfig.ini
@@ -10,8 +10,8 @@
         "ca_certificate": "ca_cert.pem",
         "ca_private_key": "ca_key.pem",
         "upload_path": "uploads",
-        "debug_mode": true,
-        "enable_docs": true
+        "debug_mode": false,
+        "enable_docs": false
     },
     "Database": {
         "db_type": "QPSQL",
