### 2023-10-27 #5497 @doumdi
@tmickleydoyle Thanks for the suggestions, they are now merged into the joss-paper branch . For the docker installation, have a look at : https://introlab.github.io/opentera/deployment/DeployWithDockerCompose.html.  Any suggestions / comments on the procedure is welcomed ! Thanks!
