### 2023-10-27 #5497 @tmickleydoyle
Thank you all for the ⚡ fast update to the install process 😲 I will check out the docker install early next week 🙇
