### 2023-10-30 #5497 @tmickleydoyle
@doumdi 🔆 Thank you for adding the instructions to install the software with Docker 🙇

@galessiorob 👋 I am 💯 to go with my checklist 🚀
