### 2023-11-06 #5497 @galessiorob
Thank you all for your work to get this paper ready to be published! Since we have two full reviews we're ready to take the next steps:
- I have removed @NoujoudNader due to lack of response (in this issue and via email)
- @doumdi we need you to deposit your software in Zenodo and post the DOI here and the most recent version. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing.

@Rocsg and @tmickleydoyle  thank you so, so much for your thorough and thoughtful reviews 🙏
