### 2023-11-07 #5497 @doumdi
@galessiorob You can cite all versions by using the DOI [10.5281/zenodo.10078846](https://zenodo.org/doi/10.5281/zenodo.10078846). This DOI represents all versions, and will always resolve to the latest one. Thank you @tmickleydoyle , @Kevin-Mattheus-Moerman ,  @Rocsg, @galessiorob ! BTW, the latest paper version is merged in the main branch.
