### 2023-11-09 #5497 @Kevin-Mattheus-Moerman
@galessiorob thanks for editing here. In the future you can use the following new features `@editorialbot create post-review checklist`. That will create this list :point_up: which will help keep track of the last steps needed before recommending acceptance. I'll take it from here now though. Thanks again.
