### 2023-11-13 #5497 @doumdi
@Kevin-Mattheus-Moerman , I have made the required changes on Zenodo (authors, licence, title, and added summary) here: https://zenodo.org/records/10078847. I am unable to check the boxes in your previous post, but I think everything is done. Please let me know if more changes are required. Thanks!
